refactor(gui-test): split test_gui_main.cpp into 10 files by functional domain
Split the monolithic 3131-line test_gui_main.cpp into focused test files:
- test_gui_shared.hpp (shared structs, extern declarations, Register declarations)
- test_gui_smoke.cpp, test_gui_interaction.cpp, test_gui_render.cpp,
  test_gui_export.cpp, test_gui_visual.cpp, test_gui_bg.cpp,
  test_gui_import_export.cpp, test_gui_perf.cpp
- test_gui_main.cpp retains main(), global state definitions, ResetTestState()

All 60/60 GUI tests pass. No test logic changes.
