+++ test/gui/test_gui_main.cpp
@@ -1,5 +1,4 @@
 #include <GLFW/glfw3.h>
-#include <stb_image.h>
 
 // clang-format off
 #ifdef _WIN32
@@ -8,20 +7,17 @@
 #endif
 // clang-format on
 
-#include <algorithm>
-#include <atomic>
 #include <chrono>
-#include <cmath>
 #include <cstdio>
 #include <cstring>
-#include <fstream>
 #include <string>
+#include <thread>
+#include <vector>
 
 #include "gui/gl_common.h"
 
 #define IMGUI_DEFINE_MATH_OPERATORS
 #include "gui/app.hpp"
-#include "gui/file_io.hpp"
 #include "gui/gl_init.h"
 #include "gui/gui_logger.hpp"
 #include "gui/log_sink.hpp"
@@ -30,89 +26,32 @@
 #include "imgui_impl_glfw.h"
 #include "imgui_impl_opengl3.h"
 #include "imgui_internal.h"
-#include "imgui_te_context.h"
 #include "imgui_te_engine.h"
 #include "imgui_te_exporters.h"
-#include "test_screenshot.hpp"
+#include "test_gui_shared.hpp"
 
-namespace gui = lumice::gui;
+// ========== Global variable definitions ==========
 
-// Shared state for screenshot tests: GL calls must happen on main thread (GuiFunc),
-// verification happens on test thread (TestFunc).
-struct ScreenshotCapture {
-  std::vector<unsigned char> pixels;
-  int width = 0;
-  int height = 0;
-  std::atomic<bool> capture_requested{ false };
-  std::atomic<bool> capture_done{ false };
+ScreenshotCapture g_capture;
+ExportTestState g_export_test;
+BgOverlayTestState g_bg_test;
+int g_core_log_level = LUMICE_LOG_INFO;
+int g_gui_log_level = LUMICE_LOG_INFO;
+bool g_enable_visible = false;
+bool g_enable_vsync = false;
+bool g_enable_frame_limit = true;
+bool g_enable_main_loop_commit = false;
+bool g_enable_log_panel = false;
+int g_dorun_delay_ms = 0;
+int g_main_loop_restart_count = 0;
+unsigned long g_main_loop_cumulative_rays = 0;
 
-  void Reset() {
-    pixels.clear();
-    width = 0;
-    height = 0;
-    capture_requested = false;
-    capture_done = false;
-  }
-};
-static ScreenshotCapture g_capture;
+// Synthetic texture for export tests
+std::vector<unsigned char> g_synth_tex;
 
-// Shared state for export tests: GL calls must happen on main thread (GuiFunc).
-struct ExportTestState {
-  std::string export_path;
-  bool upload_requested = false;
-  bool upload_done = false;
-  bool export_requested = false;
-  bool export_done = false;
-  bool export_result = false;
+// ========== Shared function definitions ==========
 
-  void Reset() {
-    export_path.clear();
-    upload_requested = false;
-    upload_done = false;
-    export_requested = false;
-    export_done = false;
-    export_result = false;
-  }
-};
-static ExportTestState g_export_test;
-
-// Shared state for background overlay tests: GL calls must happen on main thread (GuiFunc).
-struct BgOverlayTestState {
-  std::string bg_image_path;
-  bool bg_upload_requested = false;
-  bool bg_upload_done = false;
-  bool export_requested = false;
-  bool export_done = false;
-  bool export_result = false;
-  std::string export_path;
-
-  void Reset() {
-    bg_image_path.clear();
-    bg_upload_requested = false;
-    bg_upload_done = false;
-    export_requested = false;
-    export_done = false;
-    export_result = false;
-    export_path.clear();
-  }
-};
-static BgOverlayTestState g_bg_test;
-static int g_core_log_level = LUMICE_LOG_INFO;
-static int g_gui_log_level = LUMICE_LOG_INFO;
-static bool g_enable_visible = false;           // --visible: show window (enables DWM compositing path)
-static bool g_enable_vsync = false;             // --vsync: enable VSync (implies --visible)
-static bool g_enable_frame_limit = true;        // Default on: matches real app's kTargetFrameTimeMs fallback.
-                                                // Use --no-frame-limit to disable for raw throughput comparison.
-static bool g_enable_main_loop_commit = false;  // --main-loop-commit: DoRun on main thread (matches real app)
-static bool g_enable_log_panel = false;         // --log-panel: render LogPanel in main loop (matches real app)
-static int g_dorun_delay_ms = 0;                // --dorun-delay N: inject sleep after DoRun to simulate environment lag
-
-// Shared counters for main-loop-commit mode (written by main thread, read by test thread after measurement)
-static int g_main_loop_restart_count = 0;
-static unsigned long g_main_loop_cumulative_rays = 0;
-
-// Reset all global state for test isolation
-static void ResetTestState() {
+void ResetTestState() {
   // Document state (delegates to DoNew: g_state, g_preview, g_crystal_mesh_id/hash)
   gui::DoNew();
 
@@ -137,525 +76,7 @@ static void ResetTestState() {
   g_bg_test.Reset();
 }
 
-// ========== Helpers for task-test-gui-interaction ==========
-
-// Polls gui::g_state.texture_upload_count until it reaches baseline + 1 (or higher).
-// Returns true on success, false on timeout.
-//
-// IMPORTANT: caller must NOT call ctx->Yield() between reading baseline and calling this
-// helper; otherwise pending SyncFromPoller updates may be consumed concurrently, invalidating
-// the baseline. Keep the Arrange → baseline capture → Act (dirty + DoRun) → wait sequence
-// tight within a single test body.
-//
-// Uses ctx->Yield() (frame-rate limited, typically 16ms/frame) + 10ms explicit sleep between
-// checks, giving robust detection on both macOS (1ms timer) and Windows (1ms with
-// timeBeginPeriod, which the test main() sets). If a third call site appears, migrate to
-// test_helpers.hpp.
-static bool WaitForSimRestartAtLeast(ImGuiTestContext* ctx, unsigned long baseline_upload_count,
-                                     int timeout_ms = 1500) {
-  auto deadline = std::chrono::steady_clock::now() + std::chrono::milliseconds(timeout_ms);
-  while (std::chrono::steady_clock::now() < deadline) {
-    ctx->Yield();  // allow main thread to call SyncFromPoller()
-    if (gui::g_state.texture_upload_count >= baseline_upload_count + 1) {
-      return true;
-    }
-    std::this_thread::sleep_for(std::chrono::milliseconds(10));
-  }
-  return false;
-}
-
-// Triggers the two-stage delete for a referenced crystal via the Delete Crystal? modal.
-// Caller must have set state.selected_crystal to the target crystal index.
-static void ClickDeleteCrystalViaModal(ImGuiTestContext* ctx) {
-  ctx->ItemClick("**/Del##crystal");
-  ctx->Yield(2);
-  ctx->ItemClick("Delete Crystal?/Delete");
-  ctx->Yield();
-}
-
-// Opens the Delete Crystal? modal and clicks Cancel, leaving state unchanged.
-static void ClickDeleteCrystalCancelModal(ImGuiTestContext* ctx) {
-  ctx->ItemClick("**/Del##crystal");
-  ctx->Yield(2);
-  ctx->ItemClick("Delete Crystal?/Cancel");
-  ctx->Yield();
-}
-
-// Triggers the two-stage delete for a referenced filter via the Delete Filter? modal.
-// Caller must have set state.selected_filter to the target filter index.
-static void ClickDeleteFilterViaModal(ImGuiTestContext* ctx) {
-  ctx->ItemClick("**/Del##filter");
-  ctx->Yield(2);
-  ctx->ItemClick("Delete Filter?/Delete");
-  ctx->Yield();
-}
-
-// Register smoke tests
-static void RegisterSmokeTests(ImGuiTestEngine* engine) {
-  ImGuiTest* t = IM_REGISTER_TEST(engine, "gui_smoke", "default_state");
-  t->TestFunc = [](ImGuiTestContext* ctx) {
-    IM_UNUSED(ctx);
-    // Verify default state: 1 crystal, 1 renderer, 1 scattering layer
-    IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 1);
-    IM_CHECK_EQ(static_cast<int>(gui::g_state.renderers.size()), 1);
-    IM_CHECK_EQ(gui::g_state.selected_crystal, 0);
-    IM_CHECK_EQ(gui::g_state.selected_renderer, 0);
-    IM_CHECK_EQ(gui::g_state.dirty, false);
-    IM_CHECK_EQ(gui::g_state.sim_state, gui::GuiState::SimState::kIdle);
-  };
-}
-
-// P0 tests
-static void RegisterP0Tests(ImGuiTestEngine* engine) {
-  // P0: New
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p0_file", "new");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      IM_UNUSED(ctx);
-      ResetTestState();
-
-      // Modify state: add crystal, set dirty
-      gui::CrystalConfig c;
-      c.id = gui::g_state.next_crystal_id++;
-      gui::g_state.crystals.push_back(c);
-      gui::g_state.dirty = true;
-
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 2);
-      IM_CHECK_EQ(gui::g_state.dirty, true);
-
-      // DoNew resets
-      gui::DoNew();
-
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 1);
-      IM_CHECK_EQ(gui::g_state.dirty, false);
-      IM_CHECK_EQ(gui::g_state.selected_crystal, 0);
-      IM_CHECK_EQ(gui::g_preview.HasTexture(), false);
-    };
-  }
-
-  // P0: Save/Open roundtrip
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p0_file", "save_open_roundtrip");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      IM_UNUSED(ctx);
-      ResetTestState();
-
-      // Modify state
-      gui::g_state.crystals[0].type = gui::CrystalType::kPyramid;
-      gui::g_state.crystals[0].prism_h = 2.0f;
-      gui::g_state.crystals[0].upper_h = 0.3f;
-      gui::g_state.crystals[0].lower_h = 0.4f;
-      gui::g_state.sun.altitude = 30.0f;
-      gui::g_state.sim.max_hits = 12;
-
-      // Save
-      const char* tmp_path = "/tmp/lumice_gui_test.lmc";
-      bool save_ok = gui::SaveLmcFile(tmp_path, gui::g_state, gui::g_preview, false);
-      IM_CHECK(save_ok);
-
-      // Reset
-      gui::DoNew();
-      IM_CHECK_EQ(gui::g_state.crystals[0].type, gui::CrystalType::kPrism);
-
-      // Load
-      std::vector<unsigned char> tex_data;
-      int tex_w = 0;
-      int tex_h = 0;
-      bool load_ok = gui::LoadLmcFile(tmp_path, gui::g_state, tex_data, tex_w, tex_h);
-      IM_CHECK(load_ok);
-
-      // Verify roundtrip
-      IM_CHECK_EQ(gui::g_state.crystals[0].type, gui::CrystalType::kPyramid);
-      IM_CHECK_EQ(gui::g_state.crystals[0].prism_h, 2.0f);
-      IM_CHECK_EQ(gui::g_state.crystals[0].upper_h, 0.3f);
-      IM_CHECK_EQ(gui::g_state.crystals[0].lower_h, 0.4f);
-      IM_CHECK_EQ(gui::g_state.sun.altitude, 30.0f);
-      IM_CHECK_EQ(gui::g_state.sim.max_hits, 12);
-      IM_CHECK(tex_data.empty());  // save_texture=false
-
-      // Cleanup
-      std::remove(tmp_path);
-    };
-  }
-
-  // P0: Run/Stop UI state
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p0_sim", "run_stop_ui");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-
-      // Scene 1: kIdle — "Run" button should exist
-      gui::g_state.sim_state = gui::GuiState::SimState::kIdle;
-      ctx->Yield();
-      IM_CHECK(ctx->ItemExists("##TopBar/Run"));
-
-      // Scene 2: kSimulating — "Stop" button should exist
-      gui::g_state.sim_state = gui::GuiState::SimState::kSimulating;
-      ctx->Yield();
-      IM_CHECK(ctx->ItemExists("##TopBar/Stop"));
-
-      // Scene 3: kDone
-      gui::g_state.sim_state = gui::GuiState::SimState::kDone;
-      ctx->Yield();
-      IM_CHECK(ctx->ItemExists("##TopBar/Run"));  // Back to Run
-
-      // Scene 4: kModified — Revert button should appear
-      gui::g_state.sim_state = gui::GuiState::SimState::kModified;
-      ctx->Yield();
-      IM_CHECK(ctx->ItemExists("##TopBar/Revert"));
-    };
-  }
-}
-
-// P1 tests
-static void RegisterP1Tests(ImGuiTestEngine* engine) {
-  // P1: Crystal Add/Delete
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_crystal", "add_delete");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);  // Let GUI render
-
-      // Verify initial state: 1 crystal
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 1);
-
-      // Scenario A: Add a crystal, then delete the new (unreferenced) one
-      ctx->ItemClick("**/Add##crystal");
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 2);
-      IM_CHECK_EQ(gui::g_state.selected_crystal, 1);
-
-      // Delete the new crystal (not referenced by scattering) — should delete directly
-      ctx->ItemClick("**/Del##crystal");
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 1);
-
-      // Scenario B: Try to delete the referenced crystal
-      // Need at least 2 crystals for Del to be enabled
-      ctx->ItemClick("**/Add##crystal");
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 2);
-
-      // Select the first crystal (referenced by scattering)
-      gui::g_state.selected_crystal = 0;
-      ctx->Yield();
-
-      // Click Del — should open confirmation popup
-      ctx->ItemClick("**/Del##crystal");
-      ctx->Yield(2);
-
-      // Click Delete in the popup to confirm
-      ctx->ItemClick("Delete Crystal?/Delete");
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 1);
-    };
-  }
-
-  // P1: Filter Add/Delete
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_filter", "add_delete");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-
-      // Verify initial: no filters
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.filters.size()), 0);
-
-      // Switch to Filter tab
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Filter");
-      ctx->Yield();
-
-      // Add a filter
-      ctx->ItemClick("**/Add##filter");
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.filters.size()), 1);
-      IM_CHECK_EQ(gui::g_state.selected_filter, 0);
-
-      // Delete the filter (unreferenced) — direct delete
-      ctx->ItemClick("**/Del##filter");
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.filters.size()), 0);
-    };
-  }
-
-  // P1: Unsaved Changes Popup
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_file", "unsaved_popup");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-
-      // Set dirty
-      gui::g_state.dirty = true;
-      ctx->Yield();
-
-      // Click New — should trigger unsaved popup
-      ctx->ItemClick("##TopBar/New");
-      ctx->Yield(2);
-
-      // Click "Don't Save" in the popup
-      ctx->ItemClick("Unsaved Changes/Don't Save");
-
-      // Verify state was reset
-      IM_CHECK_EQ(gui::g_state.dirty, false);
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 1);
-    };
-  }
-
-  // P1: Mouse wheel over crystal preview should zoom only, not scroll parent panel
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_crystal", "preview_scroll_isolation");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-
-      // Ensure Crystal tab is selected (previous tests may have switched tabs)
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Crystal");
-      ctx->Yield(3);  // Let layout settle: tab switch → mesh rebuild → FBO render
-
-      ImGuiWindow* panel = ctx->GetWindowByRef("##LeftPanel");
-      IM_CHECK(panel != nullptr);
-
-      FILE* diag = fopen("/tmp/lumice_scroll_test.log", "w");
-
-      // Default state: scroll at top (0). Panel should have scrollable content.
-      float scroll_max = panel->ScrollMax.y;
-      if (diag) {
-        fprintf(diag, "scroll_max=%.1f content=%.1f panel_h=%.1f scroll=%.1f\n", scroll_max, panel->ContentSize.y,
-                panel->Size.y, panel->Scroll.y);
-      }
-      IM_CHECK(scroll_max > 0.0f);  // Panel content must overflow for this test
-
-      // Panel scroll is at 0 (top), so scrolling DOWN is possible.
-      // The 3D Preview is partially visible near the bottom of the visible area.
-      float scroll_before = panel->Scroll.y;
-      float zoom_before = gui::g_crystal_zoom;
-
-      // Move mouse into the visible portion of the crystal preview.
-      // Preview starts at ~(content_h - preview_size) from content top.
-      // At scroll=0, visible range is [0, panel_h]. Preview top is at
-      // content_h - scroll_max - preview_size, approximately.
-      // Use 85% of visible panel height to target the preview area.
-      ImVec2 panel_pos = panel->Pos;
-      float panel_w = panel->Size.x;
-      ImVec2 preview_center(panel_pos.x + panel_w * 0.5f, panel_pos.y + panel->Size.y * 0.85f);
-      if (diag) {
-        fprintf(diag, "mouse=(%.1f,%.1f) scroll_before=%.1f zoom_before=%.3f\n", preview_center.x, preview_center.y,
-                scroll_before, zoom_before);
-      }
-
-      ctx->MouseMoveToPos(preview_center);
-      ctx->Yield();
-
-      // Scroll DOWN (negative delta = scroll content down in ImGui)
-      ctx->MouseWheelY(-3.0f);
-      ctx->Yield(2);
-
-      float scroll_after = panel->Scroll.y;
-      float zoom_after = gui::g_crystal_zoom;
-      if (diag) {
-        fprintf(diag, "scroll: %.1f -> %.1f, zoom: %.3f -> %.3f\n", scroll_before, scroll_after, zoom_before,
-                zoom_after);
-        fclose(diag);
-      }
-
-      // Zoom should have changed (mouse was over preview)
-      IM_CHECK(zoom_after != zoom_before);
-      // Panel scroll should NOT have changed
-      IM_CHECK_EQ(scroll_after, scroll_before);
-    };
-  }
-}
-
-// P2 tests
-static void RegisterP2Tests(ImGuiTestEngine* engine) {
-  // P2: MarkDirty
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p2_state", "mark_dirty");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      IM_UNUSED(ctx);
-      ResetTestState();
-
-      IM_CHECK_EQ(gui::g_state.dirty, false);
-
-      // Mark dirty
-      gui::g_state.MarkDirty();
-      IM_CHECK_EQ(gui::g_state.dirty, true);
-
-      // DoNew resets dirty
-      gui::DoNew();
-      IM_CHECK_EQ(gui::g_state.dirty, false);
-    };
-  }
-
-  // P2: Lens switch — full-sky resets elevation/azimuth/roll
-  // Renderer invariants run every frame in RenderPreviewPanel
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p2_render", "lens_switch");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-
-      // Set non-zero view params
-      gui::g_state.renderers[0].elevation = 45.0f;
-      gui::g_state.renderers[0].azimuth = -30.0f;
-      gui::g_state.renderers[0].roll = 10.0f;
-
-      // Switch to full-sky lens type (index 4 = Dual Fisheye Equal Area)
-      gui::g_state.renderers[0].lens_type = 4;
-
-      // Renderer invariants in RenderPreviewPanel reset view angles for full-sky lenses
-      ctx->Yield(3);
-
-      IM_CHECK_EQ(gui::g_state.renderers[0].elevation, 0.0f);
-      IM_CHECK_EQ(gui::g_state.renderers[0].azimuth, 0.0f);
-      IM_CHECK_EQ(gui::g_state.renderers[0].roll, 0.0f);
-    };
-  }
-}
-
-// Aspect ratio tests
-static void RegisterAspectRatioTests(ImGuiTestEngine* engine) {
-  // Test 1: Preset switch — verify GetAspectRatio returns correct values
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "aspect_ratio", "switch");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-
-      struct TestCase {
-        gui::AspectPreset preset;
-        float expected_ratio;
-      };
-      TestCase cases[] = {
-        { gui::AspectPreset::k16x9, 16.0f / 9.0f },
-        { gui::AspectPreset::k3x2, 3.0f / 2.0f },
-        { gui::AspectPreset::k4x3, 4.0f / 3.0f },
-        { gui::AspectPreset::k1x1, 1.0f },
-      };
-
-      for (auto& tc : cases) {
-        float ratio = gui::GetAspectRatio(tc.preset);
-        IM_CHECK_LT(std::abs(ratio - tc.expected_ratio), 0.01f);
-      }
-
-      // Free and MatchBg return 0
-      IM_CHECK_EQ(gui::GetAspectRatio(gui::AspectPreset::kFree), 0.0f);
-      IM_CHECK_EQ(gui::GetAspectRatio(gui::AspectPreset::kMatchBg), 0.0f);
-    };
-  }
-
-  // Test 2: Portrait toggle
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "aspect_ratio", "portrait_toggle");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-
-      // Landscape 3:2 = 1.5
-      gui::g_state.aspect_preset = gui::AspectPreset::k3x2;
-      gui::g_state.aspect_portrait = false;
-      float ratio = gui::GetAspectRatio(gui::g_state.aspect_preset);
-      IM_CHECK_LT(std::abs(ratio - 1.5f), 0.01f);
-
-      // Portrait inverts: 1/1.5 ≈ 0.667
-      gui::g_state.aspect_portrait = true;
-      float inv_ratio = 1.0f / gui::GetAspectRatio(gui::g_state.aspect_preset);
-      IM_CHECK_LT(std::abs(inv_ratio - 2.0f / 3.0f), 0.01f);
-    };
-  }
-
-  // Test 3: Free mode — no ratio constraint
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "aspect_ratio", "free_mode");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-
-      IM_CHECK_EQ(gui::g_state.aspect_preset, gui::AspectPreset::kFree);
-      IM_CHECK_EQ(gui::g_state.aspect_portrait, false);
-      IM_CHECK_EQ(gui::GetAspectRatio(gui::AspectPreset::kFree), 0.0f);
-    };
-  }
-
-  // Test 4: .lmc roundtrip
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "aspect_ratio", "lmc_roundtrip");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      IM_UNUSED(ctx);
-      ResetTestState();
-
-      // Set 4:3 landscape
-      gui::g_state.aspect_preset = gui::AspectPreset::k4x3;
-      gui::g_state.aspect_portrait = false;
-
-      // Save
-      const char* tmp_path = "/tmp/lumice_aspect_test.lmc";
-      bool save_ok = gui::SaveLmcFile(tmp_path, gui::g_state, gui::g_preview, false);
-      IM_CHECK(save_ok);
-
-      // Reset
-      gui::DoNew();
-      IM_CHECK_EQ(gui::g_state.aspect_preset, gui::AspectPreset::kFree);
-
-      // Load
-      std::vector<unsigned char> tex_data;
-      int tex_w = 0;
-      int tex_h = 0;
-      bool load_ok = gui::LoadLmcFile(tmp_path, gui::g_state, tex_data, tex_w, tex_h);
-      IM_CHECK(load_ok);
-
-      // Verify roundtrip
-      IM_CHECK_EQ(gui::g_state.aspect_preset, gui::AspectPreset::k4x3);
-      IM_CHECK_EQ(gui::g_state.aspect_portrait, false);
-
-      std::remove(tmp_path);
-    };
-  }
-
-  // Test 5: Old .lmc compat — missing aspect fields default to Free
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "aspect_ratio", "old_lmc_compat");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      IM_UNUSED(ctx);
-      ResetTestState();
-
-      // Minimal JSON without aspect fields
-      std::string json = R"({"crystals":[],"renderers":[],"filters":[]})";
-      gui::GuiState loaded;
-      bool ok = gui::DeserializeGuiStateJson(json, loaded);
-      IM_CHECK(ok);
-      IM_CHECK_EQ(loaded.aspect_preset, gui::AspectPreset::kFree);
-      IM_CHECK_EQ(loaded.aspect_portrait, false);
-    };
-  }
-
-  // Test 6: Screen bounds — ApplyAspectRatio clamps to workarea
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "aspect_ratio", "screen_bounds");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      IM_UNUSED(ctx);
-
-      // Verify workarea is available
-      int work_x = 0;
-      int work_y = 0;
-      int work_w = 0;
-      int work_h = 0;
-      glfwGetMonitorWorkarea(glfwGetPrimaryMonitor(), &work_x, &work_y, &work_w, &work_h);
-
-      // The function should never produce sizes exceeding workarea
-      // Just verify GetAspectRatio values are sane
-      for (int i = 0; i < gui::kAspectPresetCount; i++) {
-        float r = gui::GetAspectRatio(static_cast<gui::AspectPreset>(i));
-        IM_CHECK(r >= 0.0f);
-        if (r > 0.0f) {
-          IM_CHECK(r < 100.0f);  // Sanity check
-        }
-      }
-    };
-  }
-}
-
-// Synthetic texture for export tests
-static std::vector<unsigned char> g_synth_tex;
-constexpr int kSynthTexW = 64;
-constexpr int kSynthTexH = 64;
-
-static void InitSynthTexture() {
+void InitSynthTexture() {
   if (g_synth_tex.empty()) {
     g_synth_tex.resize(kSynthTexW * kSynthTexH * 3);
     for (int y = 0; y < kSynthTexH; ++y) {
@@ -669,2162 +90,7 @@ static void InitSynthTexture() {
   }
 }
 
-static void ExportGuiFunc(ImGuiTestContext* /*ctx*/) {
-  // Upload synthetic texture on main thread when requested
-  if (g_export_test.upload_requested && !g_export_test.upload_done) {
-    InitSynthTexture();
-    gui::g_preview.UploadTexture(g_synth_tex.data(), kSynthTexW, kSynthTexH);
-    g_export_test.upload_done = true;
-  }
-  // Execute export on main thread when requested
-  if (g_export_test.export_requested && !g_export_test.export_done) {
-    g_export_test.export_result =
-        gui::ExportPreviewPng(g_export_test.export_path.c_str(), gui::g_preview, gui::g_preview_vp);
-    g_export_test.export_done = true;
-    g_export_test.export_requested = false;
-  }
-}
-
-static void RegisterExportPreviewTests(ImGuiTestEngine* engine) {
-  // Test 1: export_file_valid — export to temp file, verify stbi_load can read it
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "export", "file_valid");
-    t->GuiFunc = ExportGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      g_export_test.Reset();
-
-      // Request texture upload
-      g_export_test.upload_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_export_test.upload_done);
-      IM_CHECK(gui::g_preview.HasTexture());
-
-      // Ensure preview viewport has valid dimensions
-      ctx->Yield(2);
-      IM_CHECK(gui::g_preview_vp.vp_w > 0);
-      IM_CHECK(gui::g_preview_vp.vp_h > 0);
-
-      // Request export
-      const char* tmp_path = "/tmp/lumice_export_test.png";
-      g_export_test.export_path = tmp_path;
-      g_export_test.export_requested = true;
-      ctx->Yield(2);
-
-      IM_CHECK(g_export_test.export_done);
-      IM_CHECK(g_export_test.export_result);
-
-      // Verify file exists and is readable
-      std::vector<unsigned char> img_data;
-      int img_w = 0;
-      int img_h = 0;
-      int img_ch = 0;
-      bool loaded = lumice::test::LoadPng(tmp_path, img_data, img_w, img_h, img_ch);
-      IM_CHECK(loaded);
-      IM_CHECK(img_w > 0);
-      IM_CHECK(img_h > 0);
-
-      std::remove(tmp_path);
-    };
-  }
-
-  // Test 2: export_dimensions — verify exported image matches viewport size
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "export", "dimensions");
-    t->GuiFunc = ExportGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      g_export_test.Reset();
-
-      // Upload texture
-      g_export_test.upload_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_export_test.upload_done);
-
-      // Wait for viewport to be active
-      ctx->Yield(2);
-      int expected_w = gui::g_preview_vp.vp_w;
-      int expected_h = gui::g_preview_vp.vp_h;
-      IM_CHECK(expected_w > 0);
-      IM_CHECK(expected_h > 0);
-
-      // Export
-      const char* tmp_path = "/tmp/lumice_export_dim_test.png";
-      g_export_test.export_path = tmp_path;
-      g_export_test.export_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_export_test.export_done);
-      IM_CHECK(g_export_test.export_result);
-
-      // Verify dimensions
-      std::vector<unsigned char> img_data;
-      int img_w = 0;
-      int img_h = 0;
-      int img_ch = 0;
-      bool loaded = lumice::test::LoadPng(tmp_path, img_data, img_w, img_h, img_ch);
-      IM_CHECK(loaded);
-      IM_CHECK_EQ(img_w, expected_w);
-      IM_CHECK_EQ(img_h, expected_h);
-
-      std::remove(tmp_path);
-    };
-  }
-
-  // Test 3: export_content — verify exported image has non-zero pixels (content test)
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "export", "content");
-    t->GuiFunc = ExportGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      g_export_test.Reset();
-
-      // Upload texture
-      g_export_test.upload_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_export_test.upload_done);
-
-      ctx->Yield(2);
-
-      // Export
-      const char* tmp_path = "/tmp/lumice_export_content_test.png";
-      g_export_test.export_path = tmp_path;
-      g_export_test.export_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_export_test.export_done);
-      IM_CHECK(g_export_test.export_result);
-
-      // Verify content: load and check for non-zero pixels
-      std::vector<unsigned char> img_data;
-      int img_w = 0;
-      int img_h = 0;
-      int img_ch = 0;
-      bool loaded = lumice::test::LoadPng(tmp_path, img_data, img_w, img_h, img_ch);
-      IM_CHECK(loaded);
-      bool has_nonzero = false;
-      for (size_t i = 0; i < img_data.size() && !has_nonzero; ++i) {
-        if (img_data[i] != 0) {
-          has_nonzero = true;
-        }
-      }
-      IM_CHECK(has_nonzero);
-
-      std::remove(tmp_path);
-    };
-  }
-
-  // Test 4: export_twice — export twice consecutively, verify both succeed (FBO cleanup test)
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "export", "twice");
-    t->GuiFunc = ExportGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      g_export_test.Reset();
-
-      // Upload texture
-      g_export_test.upload_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_export_test.upload_done);
-      ctx->Yield(2);
-
-      // First export
-      const char* tmp_path1 = "/tmp/lumice_export_twice_1.png";
-      g_export_test.export_path = tmp_path1;
-      g_export_test.export_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_export_test.export_done);
-      IM_CHECK(g_export_test.export_result);
-
-      // Second export (verify FBO cleanup is correct — prev_fbo restored properly)
-      g_export_test.export_done = false;
-      g_export_test.export_result = false;
-      const char* tmp_path2 = "/tmp/lumice_export_twice_2.png";
-      g_export_test.export_path = tmp_path2;
-      g_export_test.export_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_export_test.export_done);
-      IM_CHECK(g_export_test.export_result);
-
-      // Verify both files are readable
-      std::vector<unsigned char> img1, img2;
-      int w1 = 0, h1 = 0, ch1 = 0;
-      int w2 = 0, h2 = 0, ch2 = 0;
-      IM_CHECK(lumice::test::LoadPng(tmp_path1, img1, w1, h1, ch1));
-      IM_CHECK(lumice::test::LoadPng(tmp_path2, img2, w2, h2, ch2));
-      IM_CHECK_EQ(w1, w2);
-      IM_CHECK_EQ(h1, h2);
-
-      std::remove(tmp_path1);
-      std::remove(tmp_path2);
-    };
-  }
-}
-
-// GuiFunc that renders GUI and captures crystal texture when requested
-static void ScreenshotGuiFunc(ImGuiTestContext* /*ctx*/) {
-  // Normal GUI rendering is handled by the main loop.
-  // Capture crystal texture on main thread when requested.
-  if (g_capture.capture_requested && !g_capture.capture_done) {
-    int w = gui::g_crystal_renderer.Width();
-    int h = gui::g_crystal_renderer.Height();
-    auto tex_id = static_cast<unsigned int>(gui::g_crystal_renderer.GetTextureId());
-    g_capture.pixels = lumice::test::ReadTexturePixels(tex_id, w, h);
-    g_capture.width = w;
-    g_capture.height = h;
-    g_capture.capture_done = true;
-  }
-}
-
-// Screenshot tests
-static void RegisterScreenshotTests(ImGuiTestEngine* engine) {
-  // Smoke: capture crystal preview, verify non-black
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "screenshot", "smoke");
-    t->GuiFunc = ScreenshotGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      g_capture.Reset();
-
-      // Crystal tab is the default first tab — no need to switch.
-      // Yield 2 frames: frame 1 triggers UpdateMesh (g_crystal_mesh_id == -1) + FBO render,
-      // frame 2 ensures resolve is complete.
-      ctx->Yield(2);
-
-      // Request capture on main thread
-      g_capture.capture_requested = true;
-      // Yield to let GuiFunc run and perform the capture
-      ctx->Yield(2);
-
-      IM_CHECK(g_capture.capture_done);
-      IM_CHECK_EQ(static_cast<int>(g_capture.pixels.size()), g_capture.width * g_capture.height * 4);
-
-      // Save to temp file
-      const char* tmp_path = "/tmp/lumice_crystal_test.png";
-      bool saved = lumice::test::SavePng(tmp_path, g_capture.pixels.data(), g_capture.width, g_capture.height, 4);
-      IM_CHECK(saved);
-
-      // Verify not all black: at least some non-zero pixels
-      bool has_nonzero = false;
-      for (size_t i = 0; i < g_capture.pixels.size() && !has_nonzero; ++i) {
-        if (g_capture.pixels[i] != 0) {
-          has_nonzero = true;
-        }
-      }
-      IM_CHECK(has_nonzero);
-
-      // Verify file is non-empty
-      FILE* f = fopen(tmp_path, "rb");
-      IM_CHECK(f != nullptr);
-      if (f) {
-        fseek(f, 0, SEEK_END);
-        long file_size = ftell(f);
-        fclose(f);
-        IM_CHECK(file_size > 0);
-      }
-
-      // Cleanup
-      std::remove(tmp_path);
-    };
-  }
-
-  // PSNR: capture crystal preview and compare against reference image
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "screenshot", "crystal_psnr");
-    t->GuiFunc = ScreenshotGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      g_capture.Reset();
-
-      // Crystal tab is default — yield 2 frames for mesh update + FBO render
-      ctx->Yield(2);
-
-      // Request capture on main thread
-      g_capture.capture_requested = true;
-      ctx->Yield(2);
-
-      IM_CHECK(g_capture.capture_done);
-
-      // Load reference image
-      const char* ref_path = LUMICE_TEST_REF_DIR "/crystal_prism_default.png";
-      std::vector<unsigned char> ref_data;
-      int ref_w = 0;
-      int ref_h = 0;
-      int ref_ch = 0;
-      bool loaded = lumice::test::LoadPng(ref_path, ref_data, ref_w, ref_h, ref_ch);
-      IM_CHECK(loaded);
-      IM_CHECK_EQ(ref_w, g_capture.width);
-      IM_CHECK_EQ(ref_h, g_capture.height);
-
-      // Convert capture to match reference channels if needed (RGBA→RGB or vice versa)
-      const unsigned char* cmp_data = g_capture.pixels.data();
-      std::vector<unsigned char> converted;
-      int cmp_channels = 4;  // capture is always RGBA
-      if (ref_ch == 4) {
-        cmp_channels = 4;
-      } else if (ref_ch == 3) {
-        // Strip alpha channel for comparison
-        converted = lumice::test::StripAlpha(g_capture.pixels.data(), g_capture.width, g_capture.height);
-        cmp_data = converted.data();
-        cmp_channels = 3;
-      }
-
-      // Same-platform same-hardware should produce near-identical results
-      constexpr double kPsnrThreshold = 40.0;  // dB; deterministic GL render should be >50 dB
-      double psnr = lumice::test::ComputePsnr(cmp_data, ref_data.data(), ref_w, ref_h, cmp_channels);
-      IM_CHECK(psnr >= 0.0);
-      fprintf(stderr, "[screenshot] crystal_psnr: PSNR = %.2f dB (threshold = %.1f dB)\n", psnr, kPsnrThreshold);
-      IM_CHECK(psnr > kPsnrThreshold);
-    };
-  }
-}
-
-// Visual tests: crystal preview + render preview data correctness
-static void RegisterVisualTests(ImGuiTestEngine* engine) {
-  // Crystal preview: Pyramid (HiddenLine, default style after ResetTestState)
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "visual", "crystal_pyramid");
-    t->GuiFunc = ScreenshotGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-
-      // Ensure Crystal tab is active (previous tests may have switched to Render tab)
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Crystal");
-      ctx->Yield();
-
-      // Switch to Pyramid type
-      gui::g_state.crystals[0].type = gui::CrystalType::kPyramid;
-      // upper_h/lower_h default to 0.2 after Step 1 change
-      gui::g_crystal_mesh_id = -1;  // Force mesh rebuild
-
-      // Yield 3 frames: state change → mesh rebuild → FBO render
-      ctx->Yield(3);
-
-      // Capture
-      g_capture.Reset();
-      g_capture.capture_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_capture.capture_done);
-
-      // Save to /tmp for reference image generation
-      const char* tmp_path = "/tmp/lumice_visual_crystal_pyramid.png";
-      auto rgb = lumice::test::StripAlpha(g_capture.pixels.data(), g_capture.width, g_capture.height);
-      lumice::test::SavePng(tmp_path, rgb.data(), g_capture.width, g_capture.height, 3);
-
-      // Load reference and compare
-      const char* ref_path = LUMICE_TEST_REF_DIR "/crystal_pyramid_default.png";
-      std::vector<unsigned char> ref_data;
-      int ref_w = 0, ref_h = 0, ref_ch = 0;
-      bool loaded = lumice::test::LoadPng(ref_path, ref_data, ref_w, ref_h, ref_ch);
-      if (!loaded) {
-        fprintf(stderr, "[visual] crystal_pyramid: reference not found at %s\n", ref_path);
-        fprintf(stderr, "[visual] Run test once, then copy %s to %s\n", tmp_path, ref_path);
-        IM_CHECK(loaded);
-        return;
-      }
-      IM_CHECK_EQ(ref_w, g_capture.width);
-      IM_CHECK_EQ(ref_h, g_capture.height);
-
-      constexpr double kPsnrThreshold = 40.0;
-      double psnr = lumice::test::ComputePsnr(rgb.data(), ref_data.data(), ref_w, ref_h, ref_ch);
-      IM_CHECK(psnr >= 0.0);
-      fprintf(stderr, "[visual] crystal_pyramid: PSNR = %.2f dB\n", psnr);
-      IM_CHECK(psnr > kPsnrThreshold);
-
-      std::remove(tmp_path);
-    };
-  }
-
-  // Crystal preview: Wireframe style
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "visual", "crystal_wireframe");
-    t->GuiFunc = ScreenshotGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-
-      // Ensure Crystal tab is active
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Crystal");
-      ctx->Yield();
-
-      gui::g_crystal_style = 0;  // Wireframe
-      ctx->Yield(3);
-
-      g_capture.Reset();
-      g_capture.capture_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_capture.capture_done);
-
-      const char* tmp_path = "/tmp/lumice_visual_crystal_wireframe.png";
-      auto rgb = lumice::test::StripAlpha(g_capture.pixels.data(), g_capture.width, g_capture.height);
-      lumice::test::SavePng(tmp_path, rgb.data(), g_capture.width, g_capture.height, 3);
-
-      const char* ref_path = LUMICE_TEST_REF_DIR "/crystal_wireframe.png";
-      std::vector<unsigned char> ref_data;
-      int ref_w = 0, ref_h = 0, ref_ch = 0;
-      bool loaded = lumice::test::LoadPng(ref_path, ref_data, ref_w, ref_h, ref_ch);
-      if (!loaded) {
-        fprintf(stderr, "[visual] crystal_wireframe: reference not found at %s\n", ref_path);
-        IM_CHECK(loaded);
-        return;
-      }
-      IM_CHECK_EQ(ref_w, g_capture.width);
-      IM_CHECK_EQ(ref_h, g_capture.height);
-
-      constexpr double kPsnrThreshold = 40.0;
-      double psnr = lumice::test::ComputePsnr(rgb.data(), ref_data.data(), ref_w, ref_h, ref_ch);
-      IM_CHECK(psnr >= 0.0);
-      fprintf(stderr, "[visual] crystal_wireframe: PSNR = %.2f dB\n", psnr);
-      IM_CHECK(psnr > kPsnrThreshold);
-
-      std::remove(tmp_path);
-    };
-  }
-
-  // Crystal preview: Shaded style
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "visual", "crystal_shaded");
-    t->GuiFunc = ScreenshotGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-
-      // Ensure Crystal tab is active
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Crystal");
-      ctx->Yield();
-
-      gui::g_crystal_style = 3;  // Shaded
-      ctx->Yield(3);
-
-      g_capture.Reset();
-      g_capture.capture_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_capture.capture_done);
-
-      const char* tmp_path = "/tmp/lumice_visual_crystal_shaded.png";
-      auto rgb = lumice::test::StripAlpha(g_capture.pixels.data(), g_capture.width, g_capture.height);
-      lumice::test::SavePng(tmp_path, rgb.data(), g_capture.width, g_capture.height, 3);
-
-      const char* ref_path = LUMICE_TEST_REF_DIR "/crystal_shaded.png";
-      std::vector<unsigned char> ref_data;
-      int ref_w = 0, ref_h = 0, ref_ch = 0;
-      bool loaded = lumice::test::LoadPng(ref_path, ref_data, ref_w, ref_h, ref_ch);
-      if (!loaded) {
-        fprintf(stderr, "[visual] crystal_shaded: reference not found at %s\n", ref_path);
-        IM_CHECK(loaded);
-        return;
-      }
-      IM_CHECK_EQ(ref_w, g_capture.width);
-      IM_CHECK_EQ(ref_h, g_capture.height);
-
-      constexpr double kPsnrThreshold = 40.0;
-      double psnr = lumice::test::ComputePsnr(rgb.data(), ref_data.data(), ref_w, ref_h, ref_ch);
-      IM_CHECK(psnr >= 0.0);
-      fprintf(stderr, "[visual] crystal_shaded: PSNR = %.2f dB\n", psnr);
-      IM_CHECK(psnr > kPsnrThreshold);
-
-      std::remove(tmp_path);
-    };
-  }
-
-  // Render preview: texture upload/readback data correctness
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "visual", "preview_load");
-    t->GuiFunc = [](ImGuiTestContext* /*ctx*/) {
-      // Upload texture on main thread when requested
-      if (g_capture.capture_requested && !g_capture.capture_done) {
-        gui::g_preview.UploadTexture(g_capture.pixels.data(), g_capture.width, g_capture.height);
-        g_capture.capture_done = true;
-      }
-    };
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-
-      // Create synthetic 256x256 RGB gradient
-      constexpr int kTexW = 256;
-      constexpr int kTexH = 256;
-      std::vector<unsigned char> tex_data(kTexW * kTexH * 3);
-      for (int y = 0; y < kTexH; ++y) {
-        for (int x = 0; x < kTexW; ++x) {
-          int idx = (y * kTexW + x) * 3;
-          tex_data[idx + 0] = static_cast<unsigned char>(x);      // R = x
-          tex_data[idx + 1] = static_cast<unsigned char>(y);      // G = y
-          tex_data[idx + 2] = static_cast<unsigned char>(x ^ y);  // B = x XOR y
-        }
-      }
-
-      // Request upload via GuiFunc (GL call on main thread)
-      g_capture.Reset();
-      g_capture.pixels = tex_data;
-      g_capture.width = kTexW;
-      g_capture.height = kTexH;
-      g_capture.capture_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_capture.capture_done);
-
-      // Verify state
-      IM_CHECK(gui::g_preview.HasTexture());
-      const unsigned char* readback = gui::g_preview.GetTextureData();
-      IM_CHECK(readback != nullptr);
-
-      // Exact byte comparison (UploadTexture stores CPU copy directly)
-      bool match = (memcmp(readback, tex_data.data(), tex_data.size()) == 0);
-      if (!match) {
-        // Fallback: compute PSNR
-        double psnr = lumice::test::ComputePsnr(readback, tex_data.data(), kTexW, kTexH, 3);
-        IM_CHECK(psnr >= 0.0);
-        fprintf(stderr, "[visual] preview_load: data mismatch, PSNR = %.2f dB\n", psnr);
-      }
-      IM_CHECK(match);
-    };
-  }
-
-  // Render preview: .lmc roundtrip with texture
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "visual", "preview_lmc_roundtrip");
-    t->GuiFunc = [](ImGuiTestContext* /*ctx*/) {
-      if (g_capture.capture_requested && !g_capture.capture_done) {
-        gui::g_preview.UploadTexture(g_capture.pixels.data(), g_capture.width, g_capture.height);
-        g_capture.capture_done = true;
-      }
-    };
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-
-      // Create synthetic texture
-      constexpr int kTexW = 256;
-      constexpr int kTexH = 256;
-      std::vector<unsigned char> original(kTexW * kTexH * 3);
-      for (int y = 0; y < kTexH; ++y) {
-        for (int x = 0; x < kTexW; ++x) {
-          int idx = (y * kTexW + x) * 3;
-          original[idx + 0] = static_cast<unsigned char>((x + y) % 256);
-          original[idx + 1] = static_cast<unsigned char>((x * 2) % 256);
-          original[idx + 2] = static_cast<unsigned char>((y * 3) % 256);
-        }
-      }
-
-      // Upload via GuiFunc
-      g_capture.Reset();
-      g_capture.pixels = original;
-      g_capture.width = kTexW;
-      g_capture.height = kTexH;
-      g_capture.capture_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_capture.capture_done);
-      IM_CHECK(gui::g_preview.HasTexture());
-
-      // Save with texture
-      const char* tmp_path = "/tmp/lumice_visual_roundtrip.lmc";
-      bool save_ok = gui::SaveLmcFile(tmp_path, gui::g_state, gui::g_preview, true);
-      IM_CHECK(save_ok);
-
-      // Clear
-      gui::DoNew();
-      IM_CHECK(!gui::g_preview.HasTexture());
-
-      // Load
-      std::vector<unsigned char> loaded_tex;
-      int loaded_w = 0, loaded_h = 0;
-      bool load_ok = gui::LoadLmcFile(tmp_path, gui::g_state, loaded_tex, loaded_w, loaded_h);
-      IM_CHECK(load_ok);
-      IM_CHECK_EQ(loaded_w, kTexW);
-      IM_CHECK_EQ(loaded_h, kTexH);
-      IM_CHECK(!loaded_tex.empty());
-
-      // Re-upload loaded texture via GuiFunc
-      g_capture.Reset();
-      g_capture.pixels = loaded_tex;
-      g_capture.width = loaded_w;
-      g_capture.height = loaded_h;
-      g_capture.capture_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_capture.capture_done);
-      IM_CHECK(gui::g_preview.HasTexture());
-
-      // Exact comparison: PNG is lossless, roundtrip should be identical
-      const unsigned char* readback = gui::g_preview.GetTextureData();
-      IM_CHECK(readback != nullptr);
-      size_t data_size = static_cast<size_t>(kTexW) * kTexH * 3;
-      bool match = (memcmp(readback, original.data(), data_size) == 0);
-      if (!match) {
-        double psnr = lumice::test::ComputePsnr(readback, original.data(), kTexW, kTexH, 3);
-        IM_CHECK(psnr >= 0.0);
-        fprintf(stderr, "[visual] preview_lmc_roundtrip: mismatch, PSNR = %.2f dB\n", psnr);
-        IM_CHECK(psnr >= 60.0);  // Fallback threshold
-      } else {
-        fprintf(stderr, "[visual] preview_lmc_roundtrip: exact match\n");
-      }
-
-      // Cleanup
-      std::remove(tmp_path);
-    };
-  }
-
-  // Test: UpdateCpuTextureData → Save → Load preserves correct data and dimensions.
-  // Regression test for bug where UpdateCpuTextureData did not update tex_width_/tex_height_,
-  // causing SaveLmcFile to use stale dimensions from a previous UploadTexture/UploadXyzTexture call.
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "visual", "save_texture_dimension_consistency");
-    t->GuiFunc = [](ImGuiTestContext* /*ctx*/) {
-      if (g_capture.capture_requested && !g_capture.capture_done) {
-        gui::g_preview.UploadTexture(g_capture.pixels.data(), g_capture.width, g_capture.height);
-        g_capture.capture_done = true;
-      }
-    };
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-
-      // Step 1: Upload a 256x256 texture via UploadTexture (simulates normal XYZ upload setting tex_width_/tex_height_)
-      constexpr int kOldW = 256;
-      constexpr int kOldH = 256;
-      std::vector<unsigned char> old_tex(kOldW * kOldH * 3, 0);
-      g_capture.Reset();
-      g_capture.pixels = old_tex;
-      g_capture.width = kOldW;
-      g_capture.height = kOldH;
-      g_capture.capture_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_capture.capture_done);
-      IM_CHECK_EQ(gui::g_preview.GetTextureWidth(), kOldW);
-      IM_CHECK_EQ(gui::g_preview.GetTextureHeight(), kOldH);
-
-      // Step 2: Call UpdateCpuTextureData with DIFFERENT dimensions (simulates RefreshCpuTextureForSave)
-      constexpr int kNewW = 128;
-      constexpr int kNewH = 128;
-      std::vector<unsigned char> new_tex(kNewW * kNewH * 3);
-      for (int y = 0; y < kNewH; ++y) {
-        for (int x = 0; x < kNewW; ++x) {
-          int idx = (y * kNewW + x) * 3;
-          new_tex[idx + 0] = static_cast<unsigned char>((x * 7 + y * 3) % 256);
-          new_tex[idx + 1] = static_cast<unsigned char>((x * 11 + y * 5) % 256);
-          new_tex[idx + 2] = static_cast<unsigned char>((x * 13 + y * 17) % 256);
-        }
-      }
-      gui::g_preview.UpdateCpuTextureData(new_tex.data(), kNewW, kNewH);
-
-      // Verify dimensions were updated
-      IM_CHECK_EQ(gui::g_preview.GetTextureWidth(), kNewW);
-      IM_CHECK_EQ(gui::g_preview.GetTextureHeight(), kNewH);
-
-      // Step 3: Save with texture
-      const char* tmp_path = "/tmp/lumice_save_dim_test.lmc";
-      bool save_ok = gui::SaveLmcFile(tmp_path, gui::g_state, gui::g_preview, true);
-      IM_CHECK(save_ok);
-
-      // Step 4: Load back
-      gui::GuiState loaded_state;
-      std::vector<unsigned char> loaded_tex;
-      int loaded_w = 0, loaded_h = 0;
-      bool load_ok = gui::LoadLmcFile(tmp_path, loaded_state, loaded_tex, loaded_w, loaded_h);
-      IM_CHECK(load_ok);
-
-      // Step 5: Verify loaded texture matches UpdateCpuTextureData input (not the old UploadTexture data)
-      IM_CHECK_EQ(loaded_w, kNewW);
-      IM_CHECK_EQ(loaded_h, kNewH);
-      IM_CHECK(!loaded_tex.empty());
-
-      size_t expected_size = static_cast<size_t>(kNewW) * kNewH * 3;
-      IM_CHECK_EQ(loaded_tex.size(), expected_size);
-
-      bool match = (memcmp(loaded_tex.data(), new_tex.data(), expected_size) == 0);
-      if (!match) {
-        double psnr = lumice::test::ComputePsnr(loaded_tex.data(), new_tex.data(), kNewW, kNewH, 3);
-        fprintf(stderr, "[visual] save_texture_dimension_consistency: mismatch, PSNR = %.2f dB\n", psnr);
-        IM_CHECK(psnr >= 60.0);
-      } else {
-        fprintf(stderr, "[visual] save_texture_dimension_consistency: exact match\n");
-      }
-
-      // Cleanup
-      std::remove(tmp_path);
-    };
-  }
-
-  // End-to-end: Run simulation → Save → Open → verify visual consistency.
-  // Tests the actual user workflow: the reopened file should look the same as the live display.
-  // Regression test for PostSnapshot using stale CommitConfig-time EV instead of current GUI EV.
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "visual", "save_open_visual_consistency");
-    t->GuiFunc = [](ImGuiTestContext* /*ctx*/) {
-      // Upload texture on main thread when requested
-      if (g_capture.capture_requested && !g_capture.capture_done) {
-        gui::g_preview.UploadTexture(g_capture.pixels.data(), g_capture.width, g_capture.height);
-        g_capture.capture_done = true;
-      }
-      // FBO export on main thread when requested
-      if (g_export_test.export_requested && !g_export_test.export_done) {
-        g_export_test.export_result =
-            gui::ExportPreviewPng(g_export_test.export_path, gui::g_preview, gui::g_preview_vp);
-        g_export_test.export_done = true;
-      }
-    };
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-
-      // --- Phase 1: Start simulation and wait for texture data ---
-      gui::g_server = LUMICE_CreateServer();
-      gui::g_state.sim.infinite = true;
-      gui::g_state.sim.max_hits = 8;
-      if (!gui::g_state.renderers.empty()) {
-        auto& r = gui::g_state.renderers[0];
-        r.lens_type = 0;  // Linear
-        r.fov = 120.0f;
-        r.sim_resolution_index = 0;
-        r.visible = 2;
-        r.exposure_offset = 0.0f;
-      }
-      gui::DoRun();
-
-      // Wait for first texture upload (up to 10s)
-      auto timeout = std::chrono::steady_clock::now() + std::chrono::seconds(10);
-      while (gui::g_state.texture_upload_count == 0) {
-        ctx->Yield();
-        IM_CHECK(std::chrono::steady_clock::now() < timeout);
-      }
-      // Let a few more frames accumulate for stable data
-      for (int i = 0; i < 30; i++) {
-        ctx->Yield();
-      }
-      IM_CHECK(gui::g_preview.HasTexture());
-      IM_CHECK(gui::g_preview_vp.vp_w > 0);
-      IM_CHECK(gui::g_preview_vp.vp_h > 0);
-
-      // --- Phase 2: Stop simulation, then capture (so Save uses same data as display) ---
-      gui::DoStop();
-      ctx->Yield(5);  // Let the final poller sync complete
-
-      const char* path_a = "/tmp/lumice_save_visual_A.png";
-      g_export_test.Reset();
-      g_export_test.export_path = path_a;
-      g_export_test.export_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_export_test.export_done);
-      IM_CHECK(g_export_test.export_result);
-
-      // --- Phase 3: Save ---
-      const char* lmc_path = "/tmp/lumice_save_visual_test.lmc";
-      gui::g_state.current_file_path = lmc_path;
-      gui::DoSave();
-
-      // --- Phase 4: Destroy server and Open the saved file ---
-      gui::g_server_poller.Stop();
-      LUMICE_StopServer(gui::g_server);
-      LUMICE_DestroyServer(gui::g_server);
-      gui::g_server = nullptr;
-      gui::g_state.sim_state = gui::GuiState::SimState::kIdle;
-
-      std::vector<unsigned char> tex_data;
-      int tex_w = 0, tex_h = 0;
-      bool load_ok = gui::LoadLmcFile(lmc_path, gui::g_state, tex_data, tex_w, tex_h);
-      IM_CHECK(load_ok);
-      IM_CHECK(!tex_data.empty());
-
-      // Upload loaded texture (GuiFunc will handle GL)
-      g_capture.Reset();
-      g_capture.pixels = tex_data;
-      g_capture.width = tex_w;
-      g_capture.height = tex_h;
-      g_capture.capture_requested = true;
-      ctx->Yield(2);
-
-      // --- Phase 5: Capture loaded preview (screenshot B) ---
-      const char* path_b = "/tmp/lumice_save_visual_B.png";
-      g_export_test.Reset();
-      g_export_test.export_path = path_b;
-      g_export_test.export_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_export_test.export_done);
-      IM_CHECK(g_export_test.export_result);
-
-      // --- Phase 6: Compare screenshots A and B ---
-      std::vector<unsigned char> img_a, img_b;
-      int wa = 0, ha = 0, ca = 0;
-      int wb = 0, hb = 0, cb = 0;
-      IM_CHECK(lumice::test::LoadPng(path_a, img_a, wa, ha, ca));
-      IM_CHECK(lumice::test::LoadPng(path_b, img_b, wb, hb, cb));
-      IM_CHECK_EQ(wa, wb);
-      IM_CHECK_EQ(ha, hb);
-
-      // Compare in RGB (strip alpha if needed)
-      std::vector<unsigned char> rgb_a, rgb_b;
-      int ch = std::min(ca, cb);
-      if (ca == 4) {
-        rgb_a = lumice::test::StripAlpha(img_a.data(), wa, ha);
-      } else {
-        rgb_a = img_a;
-      }
-      if (cb == 4) {
-        rgb_b = lumice::test::StripAlpha(img_b.data(), wb, hb);
-      } else {
-        rgb_b = img_b;
-      }
-
-      // PSNR threshold: the save path gets a fresh snapshot from the server (slightly more accumulated
-      // data than the poller's last upload), plus float→uint8 quantization (~48 dB theoretical max).
-      // 28 dB catches major conversion errors (wrong EV, wrong matrix) while allowing for timing noise.
-      constexpr double kPsnrThreshold = 28.0;
-      double psnr = lumice::test::ComputePsnr(rgb_a.data(), rgb_b.data(), wa, ha, 3);
-      fprintf(stderr, "[visual] save_open_visual_consistency: PSNR = %.2f dB (threshold = %.1f dB)\n", psnr,
-              kPsnrThreshold);
-      IM_CHECK(psnr > kPsnrThreshold);
-
-      // Cleanup
-      std::remove(path_a);
-      std::remove(path_b);
-      std::remove(lmc_path);
-    };
-  }
-}
-
-// Background overlay GuiFunc: handles bg texture upload, equirect upload, and FBO export on main thread.
-static void BgOverlayGuiFunc(ImGuiTestContext* /*ctx*/) {
-  // Upload equirect texture when requested (reuses g_export_test state)
-  if (g_export_test.upload_requested && !g_export_test.upload_done) {
-    InitSynthTexture();
-    gui::g_preview.UploadTexture(g_synth_tex.data(), kSynthTexW, kSynthTexH);
-    g_export_test.upload_done = true;
-  }
-  // Upload bg image from file path
-  if (g_bg_test.bg_upload_requested && !g_bg_test.bg_upload_done) {
-    int w = 0;
-    int h = 0;
-    int channels = 0;
-    unsigned char* raw = stbi_load(g_bg_test.bg_image_path.c_str(), &w, &h, &channels, 3);
-    if (raw) {
-      gui::g_preview.UploadBgTexture(raw, w, h);
-      stbi_image_free(raw);
-    }
-    g_bg_test.bg_upload_done = true;
-  }
-  // Export via FBO
-  if (g_bg_test.export_requested && !g_bg_test.export_done) {
-    g_bg_test.export_result = gui::ExportPreviewPng(g_bg_test.export_path.c_str(), gui::g_preview, gui::g_preview_vp);
-    g_bg_test.export_done = true;
-  }
-}
-
-static void RegisterBgOverlayTests(ImGuiTestEngine* engine) {
-  // Test 1: bg/aspect_landscape — load landscape bg, verify viewport aspect ≈ 2650/1580
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "bg", "aspect_landscape");
-    t->GuiFunc = BgOverlayGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      g_bg_test.Reset();
-
-      // Also need an equirect texture for preview viewport to activate
-      InitSynthTexture();
-      g_export_test.Reset();
-      g_export_test.upload_requested = true;
-
-      // Upload landscape bg image
-      g_bg_test.bg_image_path = LUMICE_TEST_REF_DIR "/bg_test_landscape.jpg";
-      g_bg_test.bg_upload_requested = true;
-      ctx->Yield(3);
-      IM_CHECK(g_bg_test.bg_upload_done);
-      IM_CHECK(gui::g_preview.HasBackground());
-
-      // Verify aspect ratio ≈ 2650/1580 ≈ 1.677
-      float bg_aspect = gui::g_preview.GetBgAspect();
-      IM_CHECK(bg_aspect > 1.5f);
-      IM_CHECK(bg_aspect < 1.8f);
-    };
-  }
-
-  // Test 2: bg/aspect_portrait — load portrait bg, verify aspect ≈ 1608/2488 ≈ 0.646
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "bg", "aspect_portrait");
-    t->GuiFunc = BgOverlayGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      g_bg_test.Reset();
-
-      g_bg_test.bg_image_path = LUMICE_TEST_REF_DIR "/bg_test_portrait.jpg";
-      g_bg_test.bg_upload_requested = true;
-      ctx->Yield(3);
-      IM_CHECK(g_bg_test.bg_upload_done);
-      IM_CHECK(gui::g_preview.HasBackground());
-
-      float bg_aspect = gui::g_preview.GetBgAspect();
-      IM_CHECK(bg_aspect > 0.5f);
-      IM_CHECK(bg_aspect < 0.8f);
-    };
-  }
-
-  // Test 3: bg/alpha_diff — alpha=0 vs alpha=1 should produce different FBO output
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "bg", "alpha_diff");
-    t->GuiFunc = BgOverlayGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      g_bg_test.Reset();
-
-      // Upload equirect texture
-      InitSynthTexture();
-      g_export_test.Reset();
-      g_export_test.upload_requested = true;
-      ctx->Yield(2);
-
-      // Upload bg image
-      g_bg_test.bg_image_path = LUMICE_TEST_REF_DIR "/bg_test_landscape.jpg";
-      g_bg_test.bg_upload_requested = true;
-      ctx->Yield(3);
-      IM_CHECK(g_bg_test.bg_upload_done);
-
-      // Enable bg overlay
-      gui::g_state.bg_show = true;
-      ctx->Yield(2);
-      IM_CHECK(gui::g_preview_vp.vp_w > 0);
-      IM_CHECK(gui::g_preview_vp.vp_h > 0);
-
-      // Export with alpha=0 (only bg visible)
-      gui::g_state.bg_alpha = 0.0f;
-      ctx->Yield(3);  // Let RenderPreviewPanel update params
-      const char* path_a0 = "/tmp/lumice_bg_alpha0.png";
-      g_bg_test.export_path = path_a0;
-      g_bg_test.export_requested = true;
-      ctx->Yield(3);
-      IM_CHECK(g_bg_test.export_done);
-      IM_CHECK(g_bg_test.export_result);
-
-      // Export with alpha=1 (only render visible)
-      g_bg_test.export_done = false;
-      g_bg_test.export_result = false;
-      g_bg_test.export_requested = false;
-      gui::g_state.bg_alpha = 1.0f;
-      ctx->Yield(3);  // Let RenderPreviewPanel update params
-      const char* path_a1 = "/tmp/lumice_bg_alpha1.png";
-      g_bg_test.export_path = path_a1;
-      g_bg_test.export_requested = true;
-      ctx->Yield(3);
-      IM_CHECK(g_bg_test.export_done);
-      IM_CHECK(g_bg_test.export_result);
-
-      // Load both and verify they differ (PSNR should be low)
-      std::vector<unsigned char> img0, img1;
-      int w0 = 0, h0 = 0, ch0 = 0;
-      int w1 = 0, h1 = 0, ch1 = 0;
-      IM_CHECK(lumice::test::LoadPng(path_a0, img0, w0, h0, ch0));
-      IM_CHECK(lumice::test::LoadPng(path_a1, img1, w1, h1, ch1));
-      IM_CHECK_EQ(w0, w1);
-      IM_CHECK_EQ(h0, h1);
-      if (w0 == w1 && h0 == h1 && ch0 == ch1) {
-        double psnr = lumice::test::ComputePsnr(img0.data(), img1.data(), w0, h0, ch0);
-        fprintf(stderr, "[bg] alpha_diff: PSNR = %.2f dB (should be < 30)\n", psnr);
-        IM_CHECK(psnr < 30.0);  // Should be quite different
-      }
-
-      std::remove(path_a0);
-      std::remove(path_a1);
-    };
-  }
-
-  // Test 4: bg/toggle_off — show=false should not affect output
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "bg", "toggle_off");
-    t->GuiFunc = BgOverlayGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      g_bg_test.Reset();
-
-      // Upload equirect texture
-      InitSynthTexture();
-      g_export_test.Reset();
-      g_export_test.upload_requested = true;
-      ctx->Yield(2);
-
-      // Upload bg but keep show=false
-      g_bg_test.bg_image_path = LUMICE_TEST_REF_DIR "/bg_test_landscape.jpg";
-      g_bg_test.bg_upload_requested = true;
-      ctx->Yield(3);
-      IM_CHECK(gui::g_preview.HasBackground());
-
-      gui::g_state.bg_show = false;
-      ctx->Yield(2);
-
-      // Verify bg_enabled is false in params
-      IM_CHECK_EQ(gui::g_preview_vp.params.bg_enabled, false);
-    };
-  }
-
-  // Test 5: bg/lmc_roundtrip — save/load preserves bg_path/bg_show/bg_alpha
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "bg", "lmc_roundtrip");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      IM_UNUSED(ctx);
-      ResetTestState();
-
-      // Set bg state
-      gui::g_state.bg_path = "/some/test/path.png";
-      gui::g_state.bg_show = true;
-      gui::g_state.bg_alpha = 0.7f;
-
-      // Save
-      const char* tmp_path = "/tmp/lumice_bg_roundtrip.lmc";
-      bool save_ok = gui::SaveLmcFile(tmp_path, gui::g_state, gui::g_preview, false);
-      IM_CHECK(save_ok);
-
-      // Reset and load
-      gui::DoNew();
-      IM_CHECK(gui::g_state.bg_path.empty());
-
-      std::vector<unsigned char> tex_data;
-      int tex_w = 0;
-      int tex_h = 0;
-      bool load_ok = gui::LoadLmcFile(tmp_path, gui::g_state, tex_data, tex_w, tex_h);
-      IM_CHECK(load_ok);
-
-      // Verify roundtrip
-      IM_CHECK_EQ(gui::g_state.bg_path, std::string("/some/test/path.png"));
-      IM_CHECK_EQ(gui::g_state.bg_show, true);
-      IM_CHECK(std::abs(gui::g_state.bg_alpha - 0.7f) < 0.01f);
-
-      std::remove(tmp_path);
-    };
-  }
-
-  // Test 6: bg/old_lmc_compat — JSON without bg fields gets defaults
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "bg", "old_lmc_compat");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      IM_UNUSED(ctx);
-      ResetTestState();
-
-      std::string json = R"({"crystals":[],"renderers":[],"filters":[]})";
-      gui::GuiState loaded;
-      bool ok = gui::DeserializeGuiStateJson(json, loaded);
-      IM_CHECK(ok);
-      IM_CHECK(loaded.bg_path.empty());
-      IM_CHECK_EQ(loaded.bg_show, false);
-      IM_CHECK(std::abs(loaded.bg_alpha - 1.0f) < 0.01f);
-    };
-  }
-
-  // Test 7: bg/match_bg_avail — Match Background availability depends on HasBackground
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "bg", "match_bg_avail");
-    t->GuiFunc = BgOverlayGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      g_bg_test.Reset();
-
-      // No bg loaded — HasBackground should be false
-      IM_CHECK_EQ(gui::g_preview.HasBackground(), false);
-
-      // Load bg
-      g_bg_test.bg_image_path = LUMICE_TEST_REF_DIR "/bg_test_landscape.jpg";
-      g_bg_test.bg_upload_requested = true;
-      ctx->Yield(3);
-      IM_CHECK(gui::g_preview.HasBackground());
-
-      // Set to kMatchBg — should work now
-      gui::g_state.aspect_preset = gui::AspectPreset::kMatchBg;
-      float bg_aspect = gui::g_preview.GetBgAspect();
-      IM_CHECK(bg_aspect > 1.0f);  // Landscape
-
-      // Clear bg — kMatchBg should not be functional
-      gui::g_preview.ClearBackground();
-      IM_CHECK_EQ(gui::g_preview.HasBackground(), false);
-    };
-  }
-
-  // Test 8: bg/contain_mode — letterbox black bars for aspect mismatch
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "bg", "contain_mode");
-    t->GuiFunc = BgOverlayGuiFunc;
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      g_bg_test.Reset();
-
-      // Upload equirect texture (needed for active viewport)
-      InitSynthTexture();
-      g_export_test.Reset();
-      g_export_test.upload_requested = true;
-      ctx->Yield(2);
-
-      // Upload landscape bg (aspect ≈ 1.677)
-      g_bg_test.bg_image_path = LUMICE_TEST_REF_DIR "/bg_test_landscape.jpg";
-      g_bg_test.bg_upload_requested = true;
-      ctx->Yield(3);
-      IM_CHECK(gui::g_preview.HasBackground());
-
-      // Enable bg with alpha=0 (full bg, no render) so we can clearly see letterbox
-      gui::g_state.bg_show = true;
-      gui::g_state.bg_alpha = 0.0f;
-      ctx->Yield(2);
-
-      IM_CHECK(gui::g_preview_vp.vp_w > 0);
-      IM_CHECK(gui::g_preview_vp.vp_h > 0);
-
-      // Export via FBO
-      const char* tmp_path = "/tmp/lumice_bg_contain.png";
-      g_bg_test.export_path = tmp_path;
-      g_bg_test.export_requested = true;
-      ctx->Yield(2);
-      IM_CHECK(g_bg_test.export_done);
-      IM_CHECK(g_bg_test.export_result);
-
-      // Load exported image
-      std::vector<unsigned char> img;
-      int img_w = 0;
-      int img_h = 0;
-      int img_ch = 0;
-      bool loaded = lumice::test::LoadPng(tmp_path, img, img_w, img_h, img_ch);
-      IM_CHECK(loaded);
-      IM_CHECK(img_w > 0);
-      IM_CHECK(img_h > 0);
-
-      // With landscape bg (aspect > 1) in a viewport that may be close to 16:9,
-      // if viewport is taller than bg (vp_aspect < bg_aspect), there should be
-      // black bars at top/bottom. Check top-left corner pixel is black (letterbox).
-      float vp_aspect = static_cast<float>(img_w) / static_cast<float>(img_h);
-      float bg_aspect = gui::g_preview.GetBgAspect();
-      fprintf(stderr, "[bg] contain_mode: vp_aspect=%.3f bg_aspect=%.3f img=%dx%d\n", vp_aspect, bg_aspect, img_w,
-              img_h);
-
-      // Verify center pixel is non-black (has bg content)
-      int cx = img_w / 2;
-      int cy = img_h / 2;
-      int center_idx = (cy * img_w + cx) * img_ch;
-      bool center_nonblack = (img[center_idx] > 5 || img[center_idx + 1] > 5 || img[center_idx + 2] > 5);
-      IM_CHECK(center_nonblack);
-
-      std::remove(tmp_path);
-    };
-  }
-}
-
-// ========== Performance Tests ==========
-
-static const char* CreatePerfConfig() {
-  // Minimal config: single prism crystal, sun at 20°, infinite rays, 1024x512 resolution
-  return R"({
-    "crystal": [{"id": 1, "type": "Prism", "height": 1.0, "ratio": {"upper": 1.0, "lower": 1.0}}],
-    "filter": [],
-    "scene": {
-      "light_source": {"type": "sun", "altitude": 20.0, "azimuth": 0, "diameter": 0.5, "spectrum": "D65"},
-      "ray_num": "infinite",
-      "max_hits": 8,
-      "scattering": [{"prob": 1.0, "entries": [{"crystal": 1, "proportion": 1.0}]}]
-    },
-    "render": [{"id": 1, "lens": {"type": "rectangular", "fov": 180.0},
-                "resolution": [1024, 512], "view": {"elevation": 0, "azimuth": 0, "roll": 0},
-                "visible": "full", "background": [0, 0, 0], "opacity": 1.0, "intensity_factor": 1.0}]
-  })";
-}
-
-static void StartPerfSimulation() {
-  gui::g_server = LUMICE_CreateServer();
-  LUMICE_SetLogLevel(gui::g_server, static_cast<LUMICE_LogLevel>(g_core_log_level));
-  gui::SetGuiLogLevel(static_cast<spdlog::level::level_enum>(g_gui_log_level));
-
-  // Set up g_state to match perf config, then use DoRun() so the server's
-  // config_manager_ is populated from the same SerializeCoreConfig path.
-  gui::g_state.sun.altitude = 20.0f;
-  gui::g_state.sun.diameter = 0.5f;
-  gui::g_state.sun.spectrum_index = 2;  // D65
-  gui::g_state.sim.infinite = true;
-  gui::g_state.sim.max_hits = 8;
-  if (!gui::g_state.renderers.empty()) {
-    auto& r = gui::g_state.renderers[0];
-    r.lens_type = 1;  // Fisheye Equal Area (matches typical manual testing)
-    r.fov = 360.0f;
-    r.sim_resolution_index = 0;  // 512 → Core resolution [1024, 512], matching CreatePerfConfig
-    r.visible = 2;               // Full
-    r.background[0] = r.background[1] = r.background[2] = 0.0f;
-    r.exposure_offset = 0.0f;
-  }
-  gui::DoRun();
-}
-
-static void StopPerfSimulation() {
-  gui::g_server_poller.Stop();
-  if (gui::g_server) {
-    LUMICE_StopServer(gui::g_server);
-    LUMICE_DestroyServer(gui::g_server);
-    gui::g_server = nullptr;
-  }
-  gui::g_state.sim_state = gui::GuiState::SimState::kIdle;
-}
-
-static void ReportPerf(const char* label, unsigned long start_rays, unsigned long end_rays, double elapsed_sec) {
-  unsigned long delta = end_rays - start_rays;
-  double rays_per_sec = elapsed_sec > 0 ? static_cast<double>(delta) / elapsed_sec : 0;
-  fprintf(stderr, "[PERF] %s: %.1f rays/sec (%lu rays in %.1fs)\n", label, rays_per_sec, delta, elapsed_sec);
-}
-
-// ========== Import/Export Tests ==========
-
-static void RegisterImportExportTests(ImGuiTestEngine* engine) {
-  // Test 1: JSON round-trip — serialize then deserialize, verify key fields match
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "import_export", "json_roundtrip");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-
-      // Set up non-default state
-      gui::g_state.sim.ray_num_millions = 2.5f;
-
-      // Serialize → deserialize
-      std::string json = gui::SerializeCoreConfig(gui::g_state);
-      IM_CHECK(!json.empty());
-
-      gui::GuiState loaded = gui::InitDefaultState();
-      bool ok = gui::DeserializeFromJson(json, loaded);
-      IM_CHECK(ok);
-
-      // Verify key fields survived round-trip
-      IM_CHECK(std::abs(loaded.sim.ray_num_millions - gui::g_state.sim.ray_num_millions) < 0.01f);
-      IM_CHECK_EQ(static_cast<int>(loaded.crystals.size()), static_cast<int>(gui::g_state.crystals.size()));
-      IM_CHECK_EQ(static_cast<int>(loaded.filters.size()), static_cast<int>(gui::g_state.filters.size()));
-      IM_CHECK_EQ(static_cast<int>(loaded.renderers.size()), static_cast<int>(gui::g_state.renderers.size()));
-    };
-  }
-
-  // Test 2: JSON file round-trip — write to file, read back, deserialize, verify
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "import_export", "json_file_roundtrip");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-
-      gui::g_state.sim.ray_num_millions = 3.0f;
-      std::string json = gui::SerializeCoreConfig(gui::g_state);
-
-      const char* tmp_path = "/tmp/lumice_config_test.json";
-      bool write_ok = gui::ExportConfigJson(tmp_path, json);
-      IM_CHECK(write_ok);
-
-      // Read back
-      std::ifstream in(tmp_path);
-      IM_CHECK(in.is_open());
-      std::string read_json((std::istreambuf_iterator<char>(in)), std::istreambuf_iterator<char>());
-      in.close();
-
-      gui::GuiState loaded = gui::InitDefaultState();
-      bool ok = gui::DeserializeFromJson(read_json, loaded);
-      IM_CHECK(ok);
-      IM_CHECK(std::abs(loaded.sim.ray_num_millions - 3.0f) < 0.01f);
-
-      std::remove(tmp_path);
-    };
-  }
-
-  // Test 3: Equirect PNG export — write synthetic data, verify file is readable
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "import_export", "equirect_export");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-
-      // Create synthetic 4x2 RGB image
-      const int w = 4;
-      const int h = 2;
-      std::vector<unsigned char> data(w * h * 3, 128);
-
-      const char* tmp_path = "/tmp/lumice_equirect_test.png";
-      bool ok = gui::ExportEquirectPng(tmp_path, data.data(), w, h);
-      IM_CHECK(ok);
-
-      // Verify file is loadable
-      std::vector<unsigned char> img_data;
-      int img_w = 0;
-      int img_h = 0;
-      int img_ch = 0;
-      bool loaded = lumice::test::LoadPng(tmp_path, img_data, img_w, img_h, img_ch);
-      IM_CHECK(loaded);
-      IM_CHECK_EQ(img_w, w);
-      IM_CHECK_EQ(img_h, h);
-
-      std::remove(tmp_path);
-    };
-  }
-}
-
-
-// ========== Calibration Tests ==========
-
-static void RegisterCalibrationTests(ImGuiTestEngine* engine) {
-  ImGuiTest* t = IM_REGISTER_TEST(engine, "calibration", "no_warning_on_startup");
-  t->TestFunc = [](ImGuiTestContext* ctx) {
-    IM_UNUSED(ctx);
-    ResetTestState();
-    gui::g_state = gui::InitDefaultState();
-
-    // Create server with DEBUG log level for diagnostics
-    gui::g_server = LUMICE_CreateServer();
-    IM_CHECK(gui::g_server != nullptr);
-    LUMICE_SetLogLevel(gui::g_server, LUMICE_LOG_DEBUG);
-    gui::SetGuiLogLevel(spdlog::level::debug);
-
-    // Set up log capture: create sink and bridge Core logs
-    auto log_sink = std::make_shared<gui::ImGuiLogSink>();
-    gui::GetGuiLogger().sinks().push_back(log_sink);
-    gui::g_imgui_log_sink = log_sink;
-    LUMICE_SetLogCallback([](LUMICE_LogLevel level, const char* /*name*/, const char* message) {
-      if (gui::g_imgui_log_sink) {
-        auto spd_level = static_cast<spdlog::level::level_enum>(level);
-        gui::g_imgui_log_sink->ReceiveExternal(spd_level, message);
-      }
-    });
-
-    // Run calibration (blocks up to 2s)
-    gui::CalibrateQualityThreshold();
-
-    // Positive assertion: stats must have data
-    LUMICE_StatsResult stats[2]{};
-    LUMICE_GetStatsResults(gui::g_server, stats, 1);
-    IM_CHECK_GT(stats[0].sim_ray_num, 0UL);
-
-    // Negative assertion: no calibration warning
-    bool found_warning = false;
-    log_sink->ForEachEntry([&](size_t, const gui::LogEntry& entry) {
-      if (entry.message.find("no data produced") != std::string::npos) {
-        found_warning = true;
-      }
-    });
-    IM_CHECK(!found_warning);
-
-    // Cleanup: destroy server BEFORE removing sink (avoid dangling writes during join)
-    LUMICE_StopServer(gui::g_server);
-    LUMICE_DestroyServer(gui::g_server);
-    gui::g_server = nullptr;
-    LUMICE_SetLogCallback(nullptr);
-    gui::g_imgui_log_sink = nullptr;
-    gui::GetGuiLogger().sinks().pop_back();
-    gui::SetGuiLogLevel(spdlog::level::warn);  // Restore default level
-  };
-}
-
-
-static void RegisterPerfTests(ImGuiTestEngine* engine) {
-  // Scenario 1: Steady-state simulation (baseline)
-  // Measures: rays/sec, texture update interval (avg/min/max), texture FPS
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "perf_test", "steady_state");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      StartPerfSimulation();
-
-      // Wait for first batch of data
-      auto timeout = std::chrono::steady_clock::now() + std::chrono::seconds(10);
-      while (gui::g_state.stats_sim_ray_num == 0) {
-        ctx->Yield();
-        if (std::chrono::steady_clock::now() > timeout) {
-          fprintf(stderr, "[PERF] ERROR: No simulation data after 10s\n");
-          break;
-        }
-      }
-
-      // Measure for 2 seconds (short enough for test engine timeout)
-      unsigned long start_rays = gui::g_state.stats_sim_ray_num;
-      auto upload_before = gui::g_state.texture_upload_count;
-      auto start_time = std::chrono::steady_clock::now();
-      auto end_time = start_time + std::chrono::seconds(2);
-
-      // Track texture update intervals
-      auto last_upload_count = upload_before;
-      auto last_upload_time = start_time;
-      std::vector<double> texture_intervals_ms;
-
-      int frame_count = 0;
-      while (std::chrono::steady_clock::now() < end_time) {
-        ctx->Yield();
-        frame_count++;
-        // Detect new texture upload
-        if (gui::g_state.texture_upload_count != last_upload_count) {
-          auto now = std::chrono::steady_clock::now();
-          double interval = std::chrono::duration<double, std::milli>(now - last_upload_time).count();
-          texture_intervals_ms.push_back(interval);
-          last_upload_count = gui::g_state.texture_upload_count;
-          last_upload_time = now;
-        }
-      }
-
-      unsigned long end_rays = gui::g_state.stats_sim_ray_num;
-      double elapsed = std::chrono::duration<double>(std::chrono::steady_clock::now() - start_time).count();
-
-      fprintf(stderr, "[PERF] steady_state: %d frames in %.1fs (%.1f main_loop_FPS)\n", frame_count, elapsed,
-              frame_count / elapsed);
-      ReportPerf("steady_state", start_rays, end_rays, elapsed);
-      double rays_per_sec = elapsed > 0 ? static_cast<double>(end_rays - start_rays) / elapsed : 0;
-      IM_CHECK_GT(rays_per_sec, 0.0);
-
-      // Report texture update frequency
-      auto texture_uploads = gui::g_state.texture_upload_count - upload_before;
-      if (!texture_intervals_ms.empty()) {
-        std::sort(texture_intervals_ms.begin(), texture_intervals_ms.end());
-        double sum = 0;
-        for (auto v : texture_intervals_ms) {
-          sum += v;
-        }
-        double avg = sum / texture_intervals_ms.size();
-        double median = texture_intervals_ms[texture_intervals_ms.size() / 2];
-        double min_val = texture_intervals_ms.front();
-        double max_val = texture_intervals_ms.back();
-        fprintf(stderr, "[PERF] steady_state: %lu texture updates in %.1fs (%.1f FPS)\n", texture_uploads, elapsed,
-                texture_uploads / elapsed);
-        fprintf(stderr, "[PERF] steady_state: texture interval avg=%.0fms median=%.0fms min=%.0fms max=%.0fms\n", avg,
-                median, min_val, max_val);
-      }
-
-      StopPerfSimulation();
-    };
-  }
-
-  // Scenario 2: Parameter drag (slider interaction during simulation)
-  // All parameter changes trigger full restart (hot-update was removed in task-52.1).
-  // Measures: rays/sec, restarts, upload ratio, per-restart ray count distribution
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "perf_test", "slider_drag");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      StartPerfSimulation();
-
-      // Wait for first batch of data
-      auto timeout = std::chrono::steady_clock::now() + std::chrono::seconds(10);
-      while (gui::g_state.stats_sim_ray_num == 0) {
-        ctx->Yield();
-        if (std::chrono::steady_clock::now() > timeout) {
-          fprintf(stderr, "[PERF] ERROR: No simulation data after 10s\n");
-          break;
-        }
-      }
-
-      // Reset main-loop-commit counters for this measurement window
-      g_main_loop_cumulative_rays = 0;
-      g_main_loop_restart_count = 0;
-
-      // Measure: sweep sun altitude over 5 seconds, matching real manual slider drag.
-      // Each frame: update altitude (continuous sweep) + set dirty.
-      // When --main-loop-commit: DoRun fires on main thread (same as real app).
-      // Otherwise: DoRun fires here on test thread (original behavior).
-      auto start_time = std::chrono::steady_clock::now();
-      auto end_time = start_time + std::chrono::seconds(5);
-      auto last_commit = start_time;
-      unsigned long cumulative_rays = 0;
-      int iteration = 0;
-      int restart_count = 0;
-      auto upload_before = gui::g_state.texture_upload_count;
-      std::vector<unsigned long> per_restart_rays;  // Ray count per restart cycle
-      std::vector<unsigned long> per_upload_rays;   // Ray count at each texture upload (measures flicker)
-      std::vector<double> first_upload_ms;          // Commit → first upload delay per restart (responsiveness)
-      auto last_upload_count = upload_before;       // Independent tracker for upload detection
-      auto last_dorun_time = start_time;            // Timestamp of most recent DoRun
-      bool waiting_first_upload = false;            // True between DoRun and first upload
-
-      auto read_server_rays = [&]() -> unsigned long {
-        if (!gui::g_server) {
-          return 0;
-        }
-        LUMICE_StatsResult stats[2]{};
-        LUMICE_GetStatsResults(gui::g_server, stats, 1);
-        return stats[0].sim_ray_num;
-      };
-
-      // Sweep sun altitude between 10° and 30° (typical manual drag range)
-      constexpr float kAltMin = 10.0f;
-      constexpr float kAltMax = 30.0f;
-
-      while (std::chrono::steady_clock::now() < end_time) {
-        // Continuous triangle wave: sweep up then down, matching smooth slider drag
-        double elapsed_sec = std::chrono::duration<double>(std::chrono::steady_clock::now() - start_time).count();
-        double phase = std::fmod(elapsed_sec, 2.0);  // 2s period (1s up + 1s down)
-        float t = (phase < 1.0) ? static_cast<float>(phase) : static_cast<float>(2.0 - phase);
-        gui::g_state.sun.altitude = kAltMin + t * (kAltMax - kAltMin);
-        gui::g_state.dirty = true;
-        iteration++;
-
-        // One yield per iteration = one frame, matching real app's per-frame commit check
-        ctx->Yield();
-
-        // Track per-upload ray counts
-        if (gui::g_state.texture_upload_count != last_upload_count) {
-          per_upload_rays.push_back(gui::g_state.stats_sim_ray_num);
-          if (waiting_first_upload) {
-            auto delay = std::chrono::duration<double, std::milli>(std::chrono::steady_clock::now() - last_dorun_time);
-            first_upload_ms.push_back(delay.count());
-            waiting_first_upload = false;
-          }
-          last_upload_count = gui::g_state.texture_upload_count;
-        }
-
-        // Throttled commit: when --main-loop-commit, the main loop handles DoRun on the main thread
-        // (matching real app behavior). Otherwise, DoRun fires here on the test thread.
-        if (!g_enable_main_loop_commit) {
-          auto now = std::chrono::steady_clock::now();
-          auto commit_elapsed = std::chrono::duration_cast<std::chrono::milliseconds>(now - last_commit).count();
-          if (commit_elapsed >= gui::kCommitIntervalMs) {
-            auto rays_this_cycle = read_server_rays();
-            per_restart_rays.push_back(rays_this_cycle);
-            cumulative_rays += rays_this_cycle;
-            gui::g_state.dirty = false;
-            gui::DoRun();
-            if (g_dorun_delay_ms > 0) {
-              std::this_thread::sleep_for(std::chrono::milliseconds(g_dorun_delay_ms));
-            }
-            last_commit = now;
-            last_dorun_time = now;
-            waiting_first_upload = true;
-            restart_count++;
-          }
-        }
-      }
-
-      if (g_enable_main_loop_commit) {
-        // Main loop tracked rays and restarts; read final cycle and combine
-        auto final_rays = read_server_rays();
-        cumulative_rays = g_main_loop_cumulative_rays + final_rays;
-        restart_count = g_main_loop_restart_count;
-      } else {
-        auto final_rays = read_server_rays();
-        per_restart_rays.push_back(final_rays);
-        cumulative_rays += final_rays;
-      }
-
-      double elapsed = std::chrono::duration<double>(std::chrono::steady_clock::now() - start_time).count();
-
-      double rays_per_sec = elapsed > 0 ? static_cast<double>(cumulative_rays) / elapsed : 0;
-      fprintf(stderr, "[PERF] slider_drag: %.1f rays/sec (%lu rays in %.1fs)\n", rays_per_sec, cumulative_rays,
-              elapsed);
-      fprintf(stderr, "[PERF] slider_drag: %d param changes, %d restarts in %.1fs\n", iteration, restart_count,
-              elapsed);
-
-      auto texture_uploads = gui::g_state.texture_upload_count - upload_before;
-      double upload_ratio = restart_count > 0 ? static_cast<double>(texture_uploads) / restart_count : 0;
-      fprintf(stderr, "[PERF] slider_drag: %lu texture uploads / %d restarts (ratio: %.2f)\n", texture_uploads,
-              restart_count, upload_ratio);
-
-      // Report commit → first upload delay (responsiveness)
-      if (!first_upload_ms.empty()) {
-        std::sort(first_upload_ms.begin(), first_upload_ms.end());
-        double fsum = 0;
-        for (auto v : first_upload_ms) {
-          fsum += v;
-        }
-        double favg = fsum / first_upload_ms.size();
-        double fmedian = first_upload_ms[first_upload_ms.size() / 2];
-        double fmin = first_upload_ms.front();
-        double fmax = first_upload_ms.back();
-        int n_miss = restart_count - static_cast<int>(first_upload_ms.size());
-        fprintf(stderr,
-                "[PERF] slider_drag: first_upload avg=%.0fms median=%.0fms min=%.0fms max=%.0fms (n=%zu, %d missed)\n",
-                favg, fmedian, fmin, fmax, first_upload_ms.size(), n_miss);
-      }
-
-      // Report per-restart ray count distribution
-      if (!per_restart_rays.empty()) {
-        std::sort(per_restart_rays.begin(), per_restart_rays.end());
-        unsigned long sum = 0;
-        for (auto v : per_restart_rays) {
-          sum += v;
-        }
-        double avg = static_cast<double>(sum) / per_restart_rays.size();
-        auto median = per_restart_rays[per_restart_rays.size() / 2];
-        auto min_val = per_restart_rays.front();
-        auto max_val = per_restart_rays.back();
-        fprintf(stderr, "[PERF] slider_drag: rays/restart avg=%.0f median=%lu min=%lu max=%lu (n=%zu)\n", avg, median,
-                min_val, max_val, per_restart_rays.size());
-      }
-
-      // Report per-upload ray count distribution and CV (measures actual texture flicker)
-      if (!per_upload_rays.empty()) {
-        std::sort(per_upload_rays.begin(), per_upload_rays.end());
-        double usum = 0;
-        for (auto v : per_upload_rays) {
-          usum += v;
-        }
-        double uavg = usum / per_upload_rays.size();
-        auto umedian = per_upload_rays[per_upload_rays.size() / 2];
-        auto umin = per_upload_rays.front();
-        auto umax = per_upload_rays.back();
-
-        // Compute CV (coefficient of variation) = stddev / mean
-        double sq_diff_sum = 0;
-        for (auto v : per_upload_rays) {
-          double diff = v - uavg;
-          sq_diff_sum += diff * diff;
-        }
-        double stddev = std::sqrt(sq_diff_sum / per_upload_rays.size());
-        double cv = uavg > 0 ? stddev / uavg * 100.0 : 0;
-
-        fprintf(stderr, "[PERF] slider_drag: upload_rays avg=%.0f median=%lu min=%lu max=%lu (n=%zu)\n", uavg, umedian,
-                umin, umax, per_upload_rays.size());
-        fprintf(stderr, "[PERF] slider_drag: upload_rays CV=%.1f%%\n", cv);
-      }
-
-      // Stop simulation BEFORE assertions: if IM_CHECK fails and aborts the test function,
-      // the server would leak and its threads would run during static destruction, crashing
-      // when accessing destroyed static locals (e.g., lens_proj_map in GetProjFunc).
-      StopPerfSimulation();
-
-      IM_CHECK_GT(rays_per_sec, 0.0);
-    };
-  }
-}
-
-// ========== task-test-gui-interaction: P1 Crystal/Filter CRUD + reference handling ==========
-
-static void RegisterP1InteractionTests(ImGuiTestEngine* engine) {
-  // p1_crystal/delete_referenced_reassigns_sole_entry
-  // Contract: deleting a crystal referenced by a sole-entry scatter layer reassigns the
-  // entry's crystal_id to the first non-deleted crystal's id (HandleDeletedCrystalRefs in
-  // panels.cpp:260-282, kReassignOnly branch).
-  // NOTE: ImGui tab selection is not in g_state; click Crystal tab explicitly because
-  // earlier tests may have switched tabs.
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_crystal", "delete_referenced_reassigns_sole_entry");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Crystal");
-      ctx->Yield(2);
-
-      // Arrange: default has 1 crystal (id=1), 1 scatter layer with sole entry → id=1
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 1);
-      IM_CHECK_EQ(gui::g_state.crystals[0].id, 1);
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.scattering[0].entries.size()), 1);
-      IM_CHECK_EQ(gui::g_state.scattering[0].entries[0].crystal_id, 1);
-
-      // Add a second crystal (id=2), unreferenced
-      ctx->ItemClick("**/Add##crystal");
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 2);
-      IM_CHECK_EQ(gui::g_state.crystals[1].id, 2);
-
-      // Act: select the first (referenced) crystal and delete via modal
-      gui::g_state.selected_crystal = 0;
-      ctx->Yield();
-      ClickDeleteCrystalViaModal(ctx);
-
-      // Assert: sole-entry layer reassigned to fallback (id=2)
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 1);
-      IM_CHECK_EQ(gui::g_state.crystals[0].id, 2);
-      IM_CHECK_EQ(gui::g_state.scattering[0].entries[0].crystal_id, 2);
-    };
-  }
-
-  // p1_crystal/delete_referenced_removes_multi_entry
-  // Contract: in a multi-entry scatter layer, deleting a referenced crystal erases matching
-  // entries (HandleDeletedCrystalRefs kRemoveOnly branch), leaving the layer's other entries
-  // intact.
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_crystal", "delete_referenced_removes_multi_entry");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Crystal");
-      ctx->Yield(2);
-
-      // Add second crystal (id=2)
-      ctx->ItemClick("**/Add##crystal");
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 2);
-      IM_CHECK_EQ(gui::g_state.crystals[1].id, 2);
-
-      // Programmatically append a second entry to the default layer referencing crystal id=2
-      gui::ScatterEntry extra;
-      extra.crystal_id = gui::g_state.crystals[1].id;  // id=2
-      extra.proportion = 50.0f;
-      gui::g_state.scattering[0].entries.push_back(extra);
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.scattering[0].entries.size()), 2);
-
-      // Act: select crystal[1] (id=2) and delete — multi-entry layer → kRemoveOnly
-      gui::g_state.selected_crystal = 1;
-      ctx->Yield();
-      ClickDeleteCrystalViaModal(ctx);
-
-      // Assert: crystal[1] removed, layer now has only the original entry (crystal_id=1)
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 1);
-      IM_CHECK_EQ(gui::g_state.crystals[0].id, 1);
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.scattering[0].entries.size()), 1);
-      IM_CHECK_EQ(gui::g_state.scattering[0].entries[0].crystal_id, 1);
-    };
-  }
-
-  // p1_crystal/delete_modal_cancel_preserves_state
-  // Contract: clicking Cancel in Delete Crystal? modal leaves state unchanged and clears
-  // g_pending_delete_crystal_idx (panels.cpp:509-511).
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_crystal", "delete_modal_cancel_preserves_state");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Crystal");
-      ctx->Yield(2);
-
-      // Need 2 crystals so Del button is enabled
-      ctx->ItemClick("**/Add##crystal");
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 2);
-
-      // Act: select referenced crystal, open modal, cancel
-      gui::g_state.selected_crystal = 0;
-      ctx->Yield();
-      ClickDeleteCrystalCancelModal(ctx);
-
-      // Assert: both crystals preserved
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 2);
-    };
-  }
-
-  // p1_crystal/delete_all_then_add_new_id
-  // Contract: next_crystal_id is monotonically increasing and never reuses deleted ids
-  // (gui_state.hpp:166, InitDefaultState starts at 1).
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_crystal", "delete_all_then_add_new_id");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Crystal");
-      ctx->Yield(2);
-
-      // Default: crystals[0].id=1, next_crystal_id=2 (after InitDefaultState)
-      IM_CHECK_EQ(gui::g_state.crystals[0].id, 1);
-      IM_CHECK_EQ(gui::g_state.next_crystal_id, 2);
-
-      // Add crystal (id=2)
-      ctx->ItemClick("**/Add##crystal");
-      IM_CHECK_EQ(gui::g_state.crystals[1].id, 2);
-      IM_CHECK_EQ(gui::g_state.next_crystal_id, 3);
-
-      // Delete the first (referenced) via modal → sole entry reassigned to id=2
-      gui::g_state.selected_crystal = 0;
-      ctx->Yield();
-      ClickDeleteCrystalViaModal(ctx);
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 1);
-      IM_CHECK_EQ(gui::g_state.crystals[0].id, 2);
-
-      // Add a new crystal — must get id=3, not reuse id=1
-      ctx->ItemClick("**/Add##crystal");
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 2);
-      IM_CHECK_EQ(gui::g_state.crystals[1].id, 3);
-      IM_CHECK_EQ(gui::g_state.next_crystal_id, 4);
-    };
-  }
-
-  // p1_filter/delete_referenced_clears_scatter_filter_id
-  // Contract: deleting a filter referenced by scatter entries clears those entries'
-  // filter_id to -1 (ClearFilterReferences in panels.cpp:308-315, called from
-  // Delete Filter? modal confirm path).
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_filter", "delete_referenced_clears_scatter_filter_id");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-
-      // Add a filter via the UI
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Filter");
-      ctx->Yield();
-      ctx->ItemClick("**/Add##filter");
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.filters.size()), 1);
-      const int filter_id = gui::g_state.filters[0].id;
-
-      // Programmatically reference this filter from the default scatter entry
-      gui::g_state.scattering[0].entries[0].filter_id = filter_id;
-      ctx->Yield();
-
-      // Act: delete via modal (referenced filter → modal path)
-      ClickDeleteFilterViaModal(ctx);
-
-      // Assert: filter removed, scatter entry's filter_id cleared
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.filters.size()), 0);
-      IM_CHECK_EQ(gui::g_state.scattering[0].entries[0].filter_id, -1);
-    };
-  }
-}
-
-// ========== task-test-gui-interaction: P1 Slider boundary tests ==========
-
-static void RegisterP1SliderBoundaryTests(ImGuiTestEngine* engine) {
-  // p1_slider/ray_count_min_max — linear scale, exact boundary values expected.
-  // SliderWithInput("Rays(M)", ...) lives in RenderSceneTab (panels.cpp:609), so we must
-  // switch to the Scene tab first. Internal widget ID is "##Rays(M)_input" via
-  // PrepareSliderLayout (panels.cpp:33).
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_slider", "ray_count_min_max");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Scene");
-      ctx->Yield(2);
-
-      IM_CHECK_EQ(gui::g_state.sim.infinite, false);  // slider active when not infinite
-
-      ctx->ItemInputValue("**/##Rays(M)_input", 0.1f);
-      ctx->Yield();
-      IM_CHECK_EQ(gui::g_state.sim.ray_num_millions, 0.1f);
-
-      ctx->ItemInputValue("**/##Rays(M)_input", 100.0f);
-      ctx->Yield();
-      IM_CHECK_EQ(gui::g_state.sim.ray_num_millions, 100.0f);
-    };
-  }
-
-  // p1_slider/ray_count_infinite_toggle — toggle preserves slider value across on/off
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_slider", "ray_count_infinite_toggle");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Scene");
-      ctx->Yield(2);
-
-      ctx->ItemInputValue("**/##Rays(M)_input", 5.0f);
-      ctx->Yield();
-      IM_CHECK_EQ(gui::g_state.sim.ray_num_millions, 5.0f);
-
-      ctx->ItemClick("**/Infinite rays");
-      ctx->Yield();
-      IM_CHECK_EQ(gui::g_state.sim.infinite, true);
-
-      ctx->ItemClick("**/Infinite rays");
-      ctx->Yield();
-      IM_CHECK_EQ(gui::g_state.sim.infinite, false);
-      // Slider value preserved across toggle
-      IM_CHECK_EQ(gui::g_state.sim.ray_num_millions, 5.0f);
-    };
-  }
-
-  // p1_slider/height_min_clamped_to_0_01 — kLog scale clamps write(0) to min
-  // SliderWithInput("Height", ..., kLog) → "##Height_slider" / "##Height_input".
-  // kLog branch uses *value = std::max(*value, min_val) to defend against 0/negative
-  // from InputFloat (panels.cpp:72).
-  // NOTE: ImGui tab selection is not part of g_state — previous tests may have switched
-  // tabs, so we must explicitly click Crystal tab here.
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_slider", "height_min_clamped_to_0_01");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Crystal");
-      ctx->Yield(2);
-
-      IM_CHECK_EQ(gui::g_state.crystals[0].type, gui::CrystalType::kPrism);
-
-      // Write 0 → expected clamp to 0.01 (slider min, kLog scale cannot reach 0)
-      ctx->ItemInputValue("**/##Height_input", 0.0f);
-      ctx->Yield();
-      IM_CHECK_GE(gui::g_state.crystals[0].height, 0.01f - 1e-6f);
-      IM_CHECK_LT(gui::g_state.crystals[0].height, 0.02f);
-    };
-  }
-
-  // p1_slider/pyramid_h1_h3_allow_zero — Upper/Lower H linear 0-1, 0 is a legal value
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_slider", "pyramid_h1_h3_allow_zero");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Crystal");
-      ctx->Yield(2);
-
-      // Switch crystal type to pyramid (Upper H / Lower H sliders appear)
-      gui::g_state.crystals[0].type = gui::CrystalType::kPyramid;
-      ctx->Yield(3);
-
-      ctx->ItemInputValue("**/##Upper H_input", 0.0f);
-      ctx->Yield();
-      IM_CHECK_EQ(gui::g_state.crystals[0].upper_h, 0.0f);
-
-      ctx->ItemInputValue("**/##Lower H_input", 0.0f);
-      ctx->Yield();
-      IM_CHECK_EQ(gui::g_state.crystals[0].lower_h, 0.0f);
-    };
-  }
-
-  // p1_slider/altitude_negative_boundaries — sun altitude ±90°
-  // SliderWithInput("Altitude", ...) is also in RenderSceneTab (panels.cpp:598)
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_slider", "altitude_negative_boundaries");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Scene");
-      ctx->Yield(2);
-
-      ctx->ItemInputValue("**/##Altitude_input", -90.0f);
-      ctx->Yield();
-      IM_CHECK_EQ(gui::g_state.sun.altitude, -90.0f);
-
-      ctx->ItemInputValue("**/##Altitude_input", 90.0f);
-      ctx->Yield();
-      IM_CHECK_EQ(gui::g_state.sun.altitude, 90.0f);
-    };
-  }
-}
-
-// ========== task-test-gui-interaction: P2 Render (lens switch / overlay) ==========
-
-static void RegisterP2InteractionRenderTests(ImGuiTestEngine* engine) {
-  // p2_render/lens_switch_fov_clamp — switching lens type clamps fov to new lens's MaxFov
-  // See app_panels.cpp:610-611: fov = std::min(fov, MaxFov(lens_type)) applied per frame.
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p2_render", "lens_switch_fov_clamp");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-
-      // Set up a full-sky lens (lens_type=4 Dual Fisheye EA) with max fov 360
-      gui::g_state.renderers[0].lens_type = 4;
-      gui::g_state.renderers[0].fov = 360.0f;
-      ctx->Yield(3);
-      IM_CHECK_EQ(gui::g_state.renderers[0].fov, 360.0f);
-
-      // Switch to Linear (lens_type=0) — its MaxFov is typically much smaller
-      gui::g_state.renderers[0].lens_type = 0;
-      ctx->Yield(3);
-
-      // fov should be clamped to Linear's max (some value < 360)
-      IM_CHECK_LT(gui::g_state.renderers[0].fov, 360.0f);
-    };
-  }
-
-  // p2_render/lens_switch_preserves_overlay — overlay flags survive lens type changes
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p2_render", "lens_switch_preserves_overlay");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-
-      gui::g_state.show_horizon = true;
-      gui::g_state.show_grid = true;
-      gui::g_state.show_sun_circles = true;
-      ctx->Yield();
-
-      gui::g_state.renderers[0].lens_type = 4;  // Dual Fisheye EA (full-sky)
-      ctx->Yield(3);
-      IM_CHECK_EQ(gui::g_state.show_horizon, true);
-      IM_CHECK_EQ(gui::g_state.show_grid, true);
-      IM_CHECK_EQ(gui::g_state.show_sun_circles, true);
-
-      gui::g_state.renderers[0].lens_type = 0;  // Linear
-      ctx->Yield(3);
-      IM_CHECK_EQ(gui::g_state.show_horizon, true);
-      IM_CHECK_EQ(gui::g_state.show_grid, true);
-      IM_CHECK_EQ(gui::g_state.show_sun_circles, true);
-    };
-  }
-}
-
-// ========== task-test-gui-interaction: P1 Running simulation restart tests ==========
-
-// These tests drive a real sim via StartPerfSimulation()/StopPerfSimulation() so that
-// DoRun() has a live g_server to commit to and SyncFromPoller() can produce texture uploads.
-// Restart detection uses the monotonic texture_upload_count as the "AtLeast +1" signal.
-
-static void RegisterP1RunningTests(ImGuiTestEngine* engine) {
-  // p1_running/crystal_change_triggers_restart (prototype — see plan.md Step 5 ROI decision)
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_running", "crystal_change_triggers_restart");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      StartPerfSimulation();
-
-      // Wait for first batch of sim data
-      auto ray_timeout = std::chrono::steady_clock::now() + std::chrono::seconds(10);
-      while (gui::g_state.stats_sim_ray_num == 0) {
-        ctx->Yield();
-        if (std::chrono::steady_clock::now() > ray_timeout) {
-          fprintf(stderr, "[TEST] crystal_change_triggers_restart: timeout waiting for first sim data\n");
-          StopPerfSimulation();
-          IM_CHECK(false);
-          return;
-        }
-      }
-
-      // Wait for first texture upload (so baseline is stable and represents a completed cycle)
-      auto upload_timeout = std::chrono::steady_clock::now() + std::chrono::seconds(5);
-      while (gui::g_state.texture_upload_count == 0) {
-        ctx->Yield();
-        if (std::chrono::steady_clock::now() > upload_timeout) {
-          break;  // proceed with baseline=0 — WaitForSimRestartAtLeast will still succeed on the first upload
-        }
-      }
-
-      // Capture baseline — must NOT yield between this and the DoRun call below
-      auto baseline = gui::g_state.texture_upload_count;
-
-      // Act: change a crystal parameter, mark dirty, commit on test thread
-      gui::g_state.crystals[0].height = 2.5f;
-      gui::g_state.dirty = true;
-      gui::DoRun();
-
-      // Assert: a new texture upload happens within timeout
-      bool ok = WaitForSimRestartAtLeast(ctx, baseline, 1500);
-
-      StopPerfSimulation();
-      IM_CHECK(ok);
-    };
-  }
-
-  // p1_running/filter_change_triggers_restart (extension)
-  // Filter changes call MarkFilterDirty() which sets intensity_locked=true to block uploads
-  // until the CommitConfig restart completes. DoRun triggers the restart which unlocks.
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_running", "filter_change_triggers_restart");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-
-      // Pre-register a filter on g_state so StartPerfSimulation's DoRun commits with it
-      gui::FilterConfig f;
-      f.id = gui::g_state.next_filter_id++;
-      f.raypath_text = "3-1-5";
-      gui::g_state.filters.push_back(f);
-      gui::g_state.selected_filter = 0;
-
-      StartPerfSimulation();
-
-      // Wait for first sim data
-      auto ray_timeout = std::chrono::steady_clock::now() + std::chrono::seconds(10);
-      while (gui::g_state.stats_sim_ray_num == 0) {
-        ctx->Yield();
-        if (std::chrono::steady_clock::now() > ray_timeout) {
-          fprintf(stderr, "[TEST] filter_change_triggers_restart: timeout waiting for first sim data\n");
-          StopPerfSimulation();
-          IM_CHECK(false);
-          return;
-        }
-      }
-
-      // Wait for first texture upload
-      auto upload_timeout = std::chrono::steady_clock::now() + std::chrono::seconds(5);
-      while (gui::g_state.texture_upload_count == 0) {
-        ctx->Yield();
-        if (std::chrono::steady_clock::now() > upload_timeout) {
-          break;
-        }
-      }
-
-      auto baseline = gui::g_state.texture_upload_count;
-
-      // Act: change filter raypath, mark filter dirty, commit
-      gui::g_state.filters[0].raypath_text = "3-1-5-7";
-      gui::g_state.MarkFilterDirty();
-      gui::DoRun();
-
-      // Filter changes may take slightly longer (intensity_locked delay); use 2000ms
-      bool ok = WaitForSimRestartAtLeast(ctx, baseline, 2000);
-
-      StopPerfSimulation();
-      IM_CHECK(ok);
-    };
-  }
-
-  // p1_running/ray_count_infinite_toggle_while_running (extension)
-  // Toggle Infinite off → fixed ray count triggers a full restart.
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p1_running", "ray_count_infinite_toggle_while_running");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      StartPerfSimulation();
-      // StartPerfSimulation sets sim.infinite=true; verify
-      IM_CHECK_EQ(gui::g_state.sim.infinite, true);
-
-      // Wait for initial data + upload
-      auto ray_timeout = std::chrono::steady_clock::now() + std::chrono::seconds(10);
-      while (gui::g_state.stats_sim_ray_num == 0) {
-        ctx->Yield();
-        if (std::chrono::steady_clock::now() > ray_timeout) {
-          fprintf(stderr, "[TEST] infinite_toggle_while_running: timeout waiting for first sim data\n");
-          StopPerfSimulation();
-          IM_CHECK(false);
-          return;
-        }
-      }
-      auto upload_timeout = std::chrono::steady_clock::now() + std::chrono::seconds(5);
-      while (gui::g_state.texture_upload_count == 0) {
-        ctx->Yield();
-        if (std::chrono::steady_clock::now() > upload_timeout) {
-          break;
-        }
-      }
-
-      // Drain any pending commits before capturing baseline (guard against race M7)
-      ctx->Yield(3);
-
-      auto baseline = gui::g_state.texture_upload_count;
-
-      // Act: switch off Infinite → fixed ray count
-      gui::g_state.sim.infinite = false;
-      gui::g_state.sim.ray_num_millions = 0.5f;  // small enough that the run can finish
-      gui::g_state.dirty = true;
-      gui::DoRun();
-
-      bool ok = WaitForSimRestartAtLeast(ctx, baseline, 2000);
-
-      StopPerfSimulation();
-      IM_CHECK(ok);
-    };
-  }
-}
-
-// ========== task-test-gui-interaction: P2 Modal interaction flows ==========
-
-static void RegisterP2InteractionModalTests(ImGuiTestEngine* engine) {
-  // p2_modal/unsaved_changes_cancel_path — clicking Cancel preserves state
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p2_modal", "unsaved_changes_cancel_path");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-      ctx->ItemClick("##LeftPanel/ConfigTabs/Crystal");
-      ctx->Yield(2);
-
-      // Add a crystal so state differs from default
-      ctx->ItemClick("**/Add##crystal");
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 2);
-      gui::g_state.dirty = true;
-      ctx->Yield();
-
-      // Click New → triggers unsaved popup
-      ctx->ItemClick("##TopBar/New");
-      ctx->Yield(2);
-
-      // Click Cancel
-      ctx->ItemClick("Unsaved Changes/Cancel");
-      ctx->Yield();
-
-      // State preserved: still 2 crystals, still dirty, pending action cleared
-      IM_CHECK_EQ(gui::g_state.dirty, true);
-      IM_CHECK_EQ(static_cast<int>(gui::g_state.crystals.size()), 2);
-      IM_CHECK_EQ(static_cast<int>(gui::g_pending_action), static_cast<int>(gui::PendingAction::kNone));
-    };
-  }
-
-  // p2_modal/unsaved_changes_save_path — clicking Save writes file, resets state, performs New
-  {
-    ImGuiTest* t = IM_REGISTER_TEST(engine, "p2_modal", "unsaved_changes_save_path");
-    t->TestFunc = [](ImGuiTestContext* ctx) {
-      ResetTestState();
-      ctx->Yield(2);
-
-      // Set current_file_path so DoSave() doesn't pop a dialog
-      const char* tmp_path = "/tmp/lumice_unsaved_save.lmc";
-      std::remove(tmp_path);  // clean up any stale file from a previous failed run
-      gui::g_state.current_file_path = tmp_path;
-      // Modify state in a verifiable way
-      gui::g_state.crystals[0].type = gui::CrystalType::kPyramid;
-      gui::g_state.crystals[0].prism_h = 3.5f;
-      gui::g_state.dirty = true;
-      ctx->Yield();
-
-      // Click New → triggers popup → Save
-      ctx->ItemClick("##TopBar/New");
-      ctx->Yield(2);
-      ctx->ItemClick("Unsaved Changes/Save");
-      ctx->Yield(2);
-
-      // After Save path: DoSave wrote the file (dirty=false), DoNew() resets state, action cleared
-      IM_CHECK_EQ(gui::g_state.dirty, false);
-      IM_CHECK_EQ(static_cast<int>(gui::g_pending_action), static_cast<int>(gui::PendingAction::kNone));
-
-      // Roundtrip check: the saved file reflects the pre-New prism_h=3.5 pyramid state
-      gui::GuiState loaded;
-      std::vector<unsigned char> tex;
-      int tw = 0;
-      int th = 0;
-      bool load_ok = gui::LoadLmcFile(tmp_path, loaded, tex, tw, th);
-      IM_CHECK(load_ok);
-      IM_CHECK_EQ(loaded.crystals[0].type, gui::CrystalType::kPyramid);
-      IM_CHECK_EQ(loaded.crystals[0].prism_h, 3.5f);
-
-      std::remove(tmp_path);
-    };
-  }
-}
+// ========== Main ==========
 
 int main(int argc, char** argv) {
   // Parse CLI arguments
